Crate bmw_util

This utility crate provides data structures and other utilities used in bmw. Currently, crate::hashtable and crate::hashset data structures are supported. Both use a crate::SlabAllocator to allocate memory. By default, the data structures use a global shared thread local slab allocator, but if multi-threaded applications are needed, a dedicated thread-safe, slab allocator may be specified. It is important to note that if the default slab allocator is used (i.e. if you just call the crate::hashtable macro without specifying a slab allocator, the values stored in the data structure will be unique to each thread.

Motivation

The advantage of these implementations is that they do not allocate memory on the heap after initialization of the data structure. So, we can create a crate::hashtable or a crate::hashset and do millions of inserts, gets, removes and iterations and no heap memory will be allocated or deallocated other than some ephemeral data structures used to do serialization. If the ‘raw’ versions of these functions are called, no heap allocations are done for serialization either. Note: there are some minor exceptions since things like the crate::Slab are returned in a Box which is stored on the heap, but those data structures are just pointers so they are small and quickly allocated and deallocated and don’t pose any concerns. Dynamic heap allocations that are long-lived can cause substantial problems and these data structures are intended to alleviate those problems. The core::ops::Drop trait is also implemented so all slabs used by the data structure are freed upon the data structure going out of scope.

Performance

These data structures will never be as fast as the native Rust data structures because they require serialization and deserialization of the entries on each operation. However, the peformance is at least in the ballpark of the standard data structures. A performance tool is included in the project in the etc directory hash_perf.

Below is the output of some runs on a linux machine. All runs use get_raw/insert_raw on small pieces of data.

First, static hashtable 1000 inserts, and no gets. Performance is fairly close.

 [2022-08-14 16:44:19.387]: (INFO) Starting tests
 [2022-08-14 16:44:19.458]: (INFO) Memory used (pre_drop) = 43.116107mb
 [2022-08-14 16:44:19.458]: (INFO) Memory Allocated (pre_drop) = 66.940703mb
 [2022-08-14 16:44:19.458]: (INFO) Memory De-allocated (pre_drop) = 23.827242mb
 [2022-08-14 16:44:19.459]: (INFO) Memory used (post_drop) = 43.011571mb
 [2022-08-14 16:44:19.459]: (INFO) Memory Allocated (post_drop) = 67.020615mb
 [2022-08-14 16:44:19.459]: (INFO) Memory De-allocated (post_drop) = 24.011692mb
 [2022-08-14 16:44:19.459]: (INFO) (StaticHash) Elapsed time = 0.51ms

With rust’s standard hashmap for comparison.

 [2022-08-14 16:45:38.840]: (INFO) Starting tests
 [2022-08-14 16:45:38.912]: (INFO) Memory used (pre_drop) = 43.062101mb
 [2022-08-14 16:45:38.912]: (INFO) Memory Allocated (pre_drop) = 66.858088mb
 [2022-08-14 16:45:38.912]: (INFO) Memory De-allocated (pre_drop) = 23.798633mb
 [2022-08-14 16:45:38.912]: (INFO) Memory used (post drop) = 43.010885mb
 [2022-08-14 16:45:38.912]: (INFO) Memory Allocated (post_drop) = 66.866024mb
 [2022-08-14 16:45:38.912]: (INFO) Memory De-allocated (post_drop) = 23.857787mb
 [2022-08-14 16:45:38.912]: (INFO) (HashMap) Elapsed time = 0.33ms

With 100 gets per insert on static hashtable.

 [2022-08-14 16:47:04.933]: (INFO) Starting tests
 [2022-08-14 16:47:05.014]: (INFO) Memory used (pre_drop) = 43.116075mb
 [2022-08-14 16:47:05.014]: (INFO) Memory Allocated (pre_drop) = 71.5672mb
 [2022-08-14 16:47:05.014]: (INFO) Memory De-allocated (pre_drop) = 28.453781mb
 [2022-08-14 16:47:05.014]: (INFO) Memory used (post_drop) = 43.011539mb
 [2022-08-14 16:47:05.014]: (INFO) Memory Allocated (post_drop) = 71.647142mb
 [2022-08-14 16:47:05.014]: (INFO) Memory De-allocated (post_drop) = 28.638261mb
 [2022-08-14 16:47:05.014]: (INFO) (StaticHash) Elapsed time = 9.49ms

With rust’s standard hashmap for comparison.

 [2022-08-14 16:47:16.677]: (INFO) Starting tests
 [2022-08-14 16:47:16.750]: (INFO) Memory used (pre_drop) = 43.062201mb
 [2022-08-14 16:47:16.750]: (INFO) Memory Allocated (pre_drop) = 66.858245mb
 [2022-08-14 16:47:16.750]: (INFO) Memory De-allocated (pre_drop) = 23.79869mb
 [2022-08-14 16:47:16.750]: (INFO) Memory used (post drop) = 43.010985mb
 [2022-08-14 16:47:16.750]: (INFO) Memory Allocated (post_drop) = 66.866181mb
 [2022-08-14 16:47:16.750]: (INFO) Memory De-allocated (post_drop) = 23.857844mb
 [2022-08-14 16:47:16.750]: (INFO) (HashMap) Elapsed time = 2.06ms

So, as seen, the performance is somewhat comparable and keep in mind, that these operations are being done in the sub millisecond time frame, so it’s very unlikely that this will ultimately impact the performance of the application unless it’s an already very optimized application that needs to be very scalable. It’s important to note that the “non-raw” versions of these operations are substantially slower, but if performance is needed in parts of the application, raw access can be used and when performance doesn’t matter as much the regular/convenient serialization functions can be used. Mix and match is possible.

Use cases

The main use case for these data structures is in server applications to avoid making dynamic heap allocations, but they also offer some other interesting properties. For instance, with the standard rust collections, the entries in the hashtable are just references so they must stay in scope while they are in the hashtable. With this implementation, that is not required. The inserted items can be dropped and they will remain in the hashtable/hashset. Also, crate::StaticHashtable and crate::StaticHashset both implement the crate::Serializable trait so they can be sent from one part of an app to another or even sent over the network.

Examples

 use bmw_err::*; // errors
 use bmw_log::*; // logging
 use bmw_util::{ctx, hashtable}; // use ctx and hashtable macro

 info!(); // log at info level in this example

 fn test() -> Result<(), Error> {
     // create a context for the hashtable
     let ctx = ctx!();

     let mut hash = hashtable!()?; // hashtable with default settings

     hash.insert(ctx, &1, &2)?; // insert 1 -> 2 (default is i32)
     assert_eq!(hash.get(ctx, &1)?.unwrap(), 2); // confirm that 1 is in the table
     hash.remove(ctx, &1)?; // remove
     assert!(hash.get(ctx, &1)?.is_none()); // confirm that 1 has been removed

     Ok(())
 }

 fn test_advanced() -> Result<(), Error> {
     // create a context for the hashtable
     let ctx = ctx!();

     // create a hashtable with maximum_entries of 10.
     let mut hash1 = hashtable!(10)?;

     for i in 0..10 {
         let key = i as u128;

         // insert with a u128 key and value of String.
         // note that anything inserted must implement the [`crate::Serializable`]
         // trait.
         hash1.insert(ctx, &key, &format!("something{}", i))?;
     }

     assert!(hash1.insert(ctx, &100u128, &"anything".to_string()).is_err());

     let mut count = 0;
     // The [`std::iter::IntoIterator`] trait is implemented, so iteration
     // is simple
     for (k,v) in &hash1 {
         info!("k={},v={}", k, v)?;
         count += 1;
     }

     assert_eq!(count, 9);

     // create a hash with a non-default maximum load capacity.
     let mut hash2 = hashtable!(10, 0.5)?;

     for i in 0..10 {
         let key = i as u128;

         // insert with a u128 key and value of String.
         // note that anything inserted must implement the [`crate::Serializable`]
         // trait.
         hash2.insert(ctx, &key, &format!("something{}", i))?;
     }

     // note that hash2 appears to behave the same way as hash1. Both allow
     // for 10 entries regardless of the load factor (hash1 has a default load factor of
     // 0.75 and hash2 has a load factor configured to 0.5). The difference is that
     // when hash2 was instantiated, an entry array of length 20 would be used, whereas
     // with hash1, an array of 14 would have been used. Both ensure that exactly max_entries
     // items can be inserted. The main considerations for the user is memory consuption and
     // performance.
     assert!(hash2.insert(ctx, &100u128, &"anything".to_string()).is_err());

     Ok(())
 }

Macros

ctx

Macro to create a context which is used by the data structures.

hashset

The crate::hashset macro is used to instantiate a crate::StaticHashset. If no parameters are specified, default values of max_entries = 1_000_000 and max_load_factor = 0.75 are used. If one parameter is specified, it will be use for max_entries. If two parameters are specified, they will be used for max_entries and max_load_factor respectively. The max_entries value is the maximum number of entries that can exist in this crate::StaticHashset. The max_load_factor is the maximum load factor that this crate::StaticHashset can support. The crate::StaticHashset will calculate the size of the entry array based on these values. With a lower load factor a larger entry array must be used to support max_entries so the lower the load factor the more resources are required.

hashset_set_raw

This macro is used to put a hashset into ‘raw’ mode. After this function is called, only the ‘raw’ functions (i.e. crate::StaticHashset::insert_raw, crate::StaticHashset::contains_raw, crate::StaticHashset::remove_raw and crate::StaticHashset::iter_raw) will return useful values.

hashtable

The crate::hashtable macro is used to instantiate a crate::StaticHashtable. If no parameters are specified, default values of max_entries = 1_000_000 and max_load_factor = 0.75 are used. If one parameter is specified, it will be use for max_entries. If two parameters are specified, they will be used for max_entries and max_load_factor respectively. The max_entries value is the maximum number of entries that can exist in this crate::StaticHashtable. The max_load_factor is the maximum load factor that this crate::StaticHashtable can support. The crate::StaticHashtable will calculate the size of the entry array based on these values. With a lower load factor a larger entry array must be used to support max_entries so the lower the load factor the more resources are required.

hashtable_set_raw

This macro is used to put a hashtable into ‘raw’ mode. After this function is called, only the ‘raw’ functions (i.e. crate::StaticHashtable::insert_raw, crate::StaticHashtable::get_raw, crate::StaticHashtable::remove_raw and crate::StaticHashtable::iter_raw) will return useful values.

init_slab_allocator

Macro to initialize the global slab allocator. It is important to note that this macro only affects the thread in which it is executed in and must be called separately in each thread that you wish to initialize the global slab allocator in. Also, the data structures, like crate::StaticHashtable and crate::StaticHashset, will initialized this slab allocator with default values if this macro is not called first. Therefore, it makes sense to call this very soon after starting a thread that will use it. The slab allocator is initialized with slab_size and slab_count parameters respecitvely. slab_size is the size in bytes of slabs. The default value is 1_024 bytes. slab_count is the number of slabs to initialize. The default value is 10_240. The defaults will be used if this macro is not called.

slab_allocator

The crate::slab_allocator macro allows for creation of a slab allocator that may be used by the data structures in this crate. If no parameters are specified, the default values of slab_size equal to 1_024 and slab_count of 10_240 are used. If one paramater is specifed, it is used as slab_count and the default is used for slab_size. If two parameters are specified, it is used as slab_count and slab_size respectively. The slab_count is the total number of slabs in this slab allocator. It is important to note that additional slabs may not be added after startup. The slab_size is the size in bytes of the slabs in this instance.

Structs

BinReader

Utility wrapper for an underlying byte Reader. Defines higher level methods to write numbers, byte vectors, hashes, etc.

BinWriter

Utility wrapper for an underlying byte Writer. Defines higher level methods to write numbers, byte vectors, hashes, etc.

Context

A context which is used in many of the methods in this crate. The reason for using the context is to avoid creating new Vectors, and other structures that require heap allocations at run time. Instead the context may be created at startup and used throughout the lifecycle of the application. The crate::Context struct may be conveniently built through the crate::ctx macro.

Slab

Struct that is used as a immutable refernce to data in a slab. See crate::SlabAllocator for further details.

SlabAllocatorBuilder

Builder struct used to build slab allocators. The build functions are generally called through the crate::slab_allocator macro to create instances of crate::SlabAllocator or through the crate::init_slab_allocator to configure the global thread local crate::SlabAllocator.

SlabAllocatorConfig

Slab Allocator configuration struct. This struct is the input to the crate::SlabAllocator::init function. The two parameters are slab_size which is the size of the slabs in bytes allocated by this crate::SlabAllocator and slab_count which is the number of slabs that can be allocated by this crate::SlabAllocator.

SlabMut

Struct that is used as a mutable refernce to data in a slab. See crate::SlabAllocator for further details.

StaticHashsetBuilder

A builder struct used to build a crate::StaticHashset. This macro is called by crate::hashset. crate::StaticHashsets are generally built through that macro.

StaticHashsetConfig

The configuration struct for a StaticHashset. This struct is passed into the crate::StaticHashsetBuilder::build function. The std::default::Default trait is implemented for this trait.

StaticHashtableBuilder

A builder struct used to build a crate::StaticHashtable. This macro is called by crate::hashtable. crate::StaticHashtables are generally built through that macro.

StaticHashtableConfig

The configuration struct for a StaticHashtable. This struct is passed into the crate::StaticHashtableBuilder::build function. The std::default::Default trait is implemented for this trait.

Traits

Reader

Reader trait used for deserializing data.

Serializable

This is the trait used by all data structures to serialize and deserialize data. Anthing stored in them must implement this trait. Commonly needed implementations are built in the ser module in this crate. These include Vec, String, integer types among other things.

SlabAllocator

This trait defines the public interface to the crate::SlabAllocator. The slab allocator is used by the other data structures in this crate to avoid dynamic heap allocations. By itself, the slab allocator is fairly simple. It only allocates and frees slabs. crate::SlabAllocator::get and crate::SlabAllocator::get_mut are also provided to obtain immutable and mutable references to a slab respectively. They only contain references to the data and not copies.

StaticHashset

The crate::StaticHashset trait defines the public interface to the static hashset. The hashset in this crate uses linear probing to handle collisions and it cannot be resized after it is intialized. Configuration of the hashset is done via the crate::StaticHashsetConfig struct. The shared implementation can be instantiated as a crate::StaticHashset through the crate::StaticHashsetBuilder::build function or as a crate::StaticHashtable through the crate::StaticHashtableBuilder::build function. Although there is a different interface for each, they are very similar and share most of the implementation code. In most cases, the data structures in this crate should be instantiated through the macros, but they can also be instantiated through the builder structs as well. See crate::hashset.

StaticHashtable

The crate::StaticHashtable trait defines the public interface to the static hashtable. The hashtable in this crate uses linear probing to handle collisions and it cannot be resized after it is intialized. Configuration of the hashtable is done via the crate::StaticHashtableConfig struct. The shared implementation can be instantiated as a crate::StaticHashtable through the crate::StaticHashtableBuilder::build function or as a crate::StaticHashset through the crate::StaticHashsetBuilder::build function. Although there is a different interface for each, they are very similar and share most of the implementation code. In most cases, the data structures in this crate should be instantiated through the macros, but they can also be instantiated through the builder structs as well. See crate::hashtable.

StaticList

TODO: not implemented

StaticQueue

TODO: not implemented

ThreadPool

TODO: not implemented

Writer

Writer trait used to serializing data.

Functions

deserialize

Deserializes a Serializable from any std::io::Read implementation.

serialize

Serializes a Serializable into any std::io::Write implementation.