Macro bmw_util::array_list

macro_rules! array_list {
    ( $size:expr, $default:expr ) => { ... };
}

The crate::array_list macro builds an crate::ArrayList in the form of an impl SortableList.

Input Parameters

size (usize) (required) - the size of the array

default (bmw_ser::Serializable) (required) - a reference to the value to initialize the array with

Return

Returns crate::SortableList on success and a bmw_err::Error on failure.

Errors

bmw_err::ErrKind::IllegalArgument - if the size is 0.

Examples

 // create an array_list and iterate through it
 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::*;

 info!();

 fn main() -> Result<(), Error> {
         let mut arr = array_list!(10, &0)?;
         for _ in 0..10 {
                 arr.push(0)?;
         }

         info!("arr = {:?}", arr)?;

         for x in arr.iter() {
                 assert_eq!(x, 0);
         }

         Ok(())
 }

 // sort an array list
 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::*;

 fn main() -> Result<(), Error> {
     // create two array lists
     let mut arr1 = array_list!(10, &0)?;
     let mut arr2 = array_list!(10, &0)?;

     // add 0..10 to the list
     for i in 0..10 {
         arr1.push(i)?;
     }

     assert!(arr1.push(0).is_err()); // it's full

     // add 10..0 to the list
     for i in (0..10).rev() {
         arr2.push(i)?;
     }

     assert!(arr2.push(0).is_err()); // it's full

     // sort the arrays using the unstable and stable sort functions (underlying rust fns used)
     arr1.sort_unstable()?;
     arr2.sort()?;

     // ensure they are equal after the sorting takes place
     assert!(list_eq!(arr1, arr2));

     Ok(())
 }