Macro bmw_util::block_on

macro_rules! block_on {
    ($res:expr) => { ... };
}

This macro causes the current thread to block until the specified thread execution completes. Upon completion, the crate::PoolResult is returned.

Input Parameters

thread_pool_handle (crate::ThreadPoolHandle) (required) - The crate::ThreadPoolHandle to block on.

Return

Returns crate::PoolResult which contains the Ok or Err value returned by this task.

Errors

bmw_err::ErrKind::ThreadPanic - if the underlying task results in a thread panic. This error is returned in the crate::PoolResult::Err variant.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::*;

 info!();

 fn main() -> Result<(), Error> {
         let mut tp = thread_pool!()?;

         tp.set_on_panic(move |id, e| -> Result<(), Error> {
             let e = e.downcast_ref::<&str>().unwrap_or(&"unknown panic type");
             info!("PANIC: id={},e={}", id, e)?;
             Ok(())
         })?;

         tp.start()?;

         let tph = execute!(tp, {
             info!("executing a task in another thread!")?;
             Ok(101)
         })?;

         let res = block_on!(tph);
         assert_eq!(res, PoolResult::Ok(101));

         Ok(())
 }