Macro bmw_util::hashset

macro_rules! hashset {
    () => { ... };
    ($max_entries:expr) => { ... };
    ($max_entries:expr, $max_load_factor:expr) => { ... };
    ($max_entries:expr, $max_load_factor:expr, $slab_allocator:expr) => { ... };
}

The crate::hashset macro is used to instantiate a crate::StaticHashset. If no parameters are specified, default values of max_entries = 1_000_000 and max_load_factor = 0.75 are used. If one parameter is specified, it will be use for max_entries. If two parameters are specified, they will be used for max_entries and max_load_factor respectively. The max_entries value is the maximum number of entries that can exist in this crate::StaticHashset. The max_load_factor is the maximum load factor that this crate::StaticHashset can support. The crate::StaticHashset will calculate the size of the entry array based on these values. With a lower load factor a larger entry array must be used to support max_entries so the lower the load factor the more resources are required.

Errors

On success crate::hashset will return a crate::StaticHashset. On failure, a bmw_err::Error is returned.

bmw_err::ErrorKind::IllegalArgument if the max_load_factor is greater than 1 or less than or equal to 0 or max_entries if is 0.

Examples

 use bmw_err::*;
 use bmw_util::{hashset, init_slab_allocator, slab_allocator, SlabAllocatorConfig};

 fn main() -> Result<(), Error> {
     // initialize the default global thread local slab allocator for this thread.
     init_slab_allocator!(1_000_000, 64);

     // intialize an additional slab allocator to be used by hashtabl4 below.
     // The other hashsets will share the default global thread local slab allocator.
     let slabs = slab_allocator!(100_000, 256)?;

     // all default values
     let mut hashset1 = hashset!()?;

     // max_enries of 15,000 with default values for the rest
     let mut hashset2 = hashset!(15_000)?;

     // max_entries of 13,000 and max_load_factor of 0.95
     let mut hashset3 = hashset!(13_000, 0.95)?;

     // max_entries of 20,000, max_load_factor of 0.9 and an owned/dedicated slab allocator
     let mut hashset4 = hashset!(20_000, 0.9, slabs)?;

     hashset1.insert(&1)?;
     hashset2.insert(&2)?;
     hashset3.insert(&3)?;
     hashset4.insert(&4)?;
     

     Ok(())
 }