Macro bmw_util::hashtable

macro_rules! hashtable {
    ($($config:tt)*) => { ... };
}

The crate::hashtable macro builds a crate::Hashtable with the specified configuration.

Input Parameters

MaxEntries (usize) (optional) - The maximum number of entries that can be in this hashtable at any given time. The default value is 1_000.

MaxLoadFactor (usize) (optional) - The maximum load factor of the hashtable. The hashtable is an array based hashtable and it has a fixed size. Once the load factor is reached, insertions will return an error. The hashtable uses linear probing to handle collisions. The max_load_factor makes sure no additional insertions occur at a given ratio of entries to capacity in the array. Note that MaxEntries can always be inserted, it’s the capacity of the array that becomes larger as this ratio goes down. So if 100 MaxEntries are specified and the MaxLoadFactor is 0.5, a 200 slot array will be used and 100 entries will be allowed. The default MaxLoadFactor is 0.7.

GlobalSlabAllocator (bool) (optional) - If true, the crate::global_slab_allocator is used instead of using an internally built slab allocator. The global slab allocator is thread_local and the returned value cannot be passed to other threads. The default value is true.

SlabSize (usize) (optional) - The size of slabs for the crate::SlabAllocator associated with this crate::Hashtable. This option is only allowed if GlobalSlabAllocator is false.

SlabCount (usize) (optional) - The count of slabs. This option is only allowed if GlobalSlabAllocator is false.

Returns

A Ok(impl Hashtable<K, V>) on success or a [bmw_err::Error] on failure.

Errors

[bmw_err::ErrKind::Configuration] - If any values are specified other than the allowed values mentioned above or if there are any duplicate parameters specified.

[bmw_err::ErrKind::Configuration] - If only one of SlabSize and SlabCount are specified.

[bmw_err::ErrKind::Configuration] - If GlobalSlabAllocator is true and SlabSize or SlabCount are specified.

[bmw_err::ErrKind::Configuration] - If GlobalSlabAllocator is false and SlabSize or SlabCount are not specified.

[bmw_err::ErrKind::IllegalArgument] - If the parameters specified for the SlabSize or SlabCount are not valid. See crate::SlabAllocator.

Examples

 use bmw_util::*;
 use bmw_log::*;             
 use bmw_err::*;             

 fn main() -> Result<(), Error> {
         // create a hashtable with the specified parameters
         let mut hashtable = hashtable!(
                 MaxEntries(1_000),  
                 MaxLoadFactor(0.9),                 
                 GlobalSlabAllocator(false),         
                 SlabSize(100),      
                 SlabCount(100)
         )?;         
                     
         // do an insert, rust will figure out what type is being inserted
         hashtable.insert(&1u32, &10u128)?;
                     
         // assert that the entry was inserted
         assert_eq!(hashtable.get(&1u32)?, Some(10u128));

         // create another hashtable with defaults, this time the global slab allocator will be
         // used. Since we did not initialize it default values will be used.
         let mut hashtable = hashtable!()?;

         // do an insert, rust will figure out what type is being inserted
         hashtable.insert(&1u8, &100u16)?;

         // assert that the entry was inserted
         assert_eq!(hashtable.get(&1u8)?, Some(100u16));

         // assert that this entry is not in the table
         assert_eq!(hashtable.get(&2u8)?, None);

         Ok(())
 }