Macro bmw_util::hashtable_box

macro_rules! hashtable_box {
    ( $( $config:expr ),* ) => { ... };
}

The crate::hashtable_box macro builds a crate::Hashtable with the specified configuration and optionally the specified crate::SlabAllocator. The only difference between this macro and the crate::hashtable macro is that the returned hashtable is inserted into a Box. Specifically, the return type is a Box<dyn Hashtable>. See crate::hashtable for further details.