Macro bmw_util::init_slab_allocator

macro_rules! init_slab_allocator {
    ($slab_size:expr, $slab_count:expr) => { ... };
}

Macro to initialize the global slab allocator. It is important to note that this macro only affects the thread in which it is executed in and must be called separately in each thread that you wish to initialize the global slab allocator in. Also, the data structures, like crate::StaticHashtable and crate::StaticHashset, will initialized this slab allocator with default values if this macro is not called first. Therefore, it makes sense to call this very soon after starting a thread that will use it. The slab allocator is initialized with slab_size and slab_count parameters respecitvely. slab_size is the size in bytes of slabs. The default value is 1_024 bytes. slab_count is the number of slabs to initialize. The default value is 10_240. The defaults will be used if this macro is not called.

Errors

init_slab_allocator returns a Ok(()) on success or returns an bmw_err::Error if an error occurs.

bmw_err::ErrorKind::IllegalArgument if the slab_size is less than 48 bytes or if slab_count is equal to 0.

bmw_err::ErrorKind::IllegalState if the thread local global slab allocator for this thread has already been initialized.

Examples

 use bmw_err::*;
 use bmw_util::{ctx, hashtable, init_slab_allocator, SlabAllocatorConfig};

 fn main() -> Result<(), Error> {
     // create a context
     let ctx = ctx!();

     // initialize the slab allocator for this thread with 25_000 128 byte entries
     init_slab_allocator!(128, 25_000);

     // The hashtable will use this default thread local slab allocator
     let mut hashtable = hashtable!()?;

     hashtable.insert(ctx, &1, &2)?;
     // ...

     Ok(())
 }

See crate::slab_allocator for details on how to create a new slab allocator as opposed to the global thread local slab allocator.