Macro bmw_util::list

macro_rules! list {
    ( $( $x:expr ),* ) => { ... };
}

The crate::list macro is used to create implementations of crate::List. This macro uses the global slab allocator. To use an internal slab allocator, see crate::UtilBuilder::build_list. This macro has the same syntax as the std::vec! macro. Note that this macro and the builder function both return an implementation of the crate::SortableList trait that uses a linked-list like implementation.

Examples

 // create a list and iterate through it
 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
     let list = list![1, 2, 3, 4];

     info!("list={:?}", list)?;

     let mut i = 1;

     for x in list.iter() {
         assert_eq!(x, i);
         i += 1;
     }

     assert!(list_eq!(list, list![1, 2, 3, 4]));

     Ok(())
 }

 // sort a list
 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::*;

 fn main() -> Result<(), Error> {
     // create two lists
     let mut list1 = list![];
     let mut list2 = list![];

     // add 0..10 to the list
     for i in 0..10 {
         list1.push(i)?;
     }

     // add 10..0 to the list
     for i in (0..10).rev() {
         list2.push(i)?;
     }

     // sort the lists using the unstable and stable sort functions (underlying rust fns used)
     list1.sort_unstable()?;
     list2.sort()?;

     // ensure they are equal after the sorting takes place
     assert!(list_eq!(list1, list2));

     Ok(())
 }