Macro bmw_util::log_init

macro_rules! log_init {
    () => { ... };
    ($config:expr) => { ... };
}

Initialize the global log. This macro takes a single parameter, if none are specified, the default crate::LogConfig is used. Note that if this macro is not called before logging occurs, the default configuration is used. After either this macro is called or the default is set via another logging macro, calling this macro again will result in an error. It usually makes sense to initilize this macro very early in the startup of an application so that no unanticipated logging occurs before this macro is called by mistake.

Examples

 use bmw_err::Error;
 use bmw_log::*;
 use bmw_log::LogConfigOption::*;
 use std::path::PathBuf;

 debug!();

 fn main() -> Result<(), Error> {
     let main_log_path = "./main.log";
     log_init!(LogConfig {
         show_bt: ShowBt(false),
         show_millis: ShowMillis(false),
         // comment this line out to avoid tests actually creating this file
         // file_path: FilePath(Some(PathBuf::from(main_log_path))),
         ..Default::default()
     });

     info!("Startup complete!")?;

     Ok(())
 }

Or without calling crate::log_init…

 use bmw_err::Error;
 use bmw_log::*;
 use bmw_log::LogConfigOption::*;

 debug!();

 fn main() -> Result<(), Error> {
     info!("Startup complete!")?;
     
     Ok(())
 }

Note that in the last example, the default crate::LogConfig will be used.