Macro bmw_util::pattern

macro_rules! pattern {
    ( $( $pattern_items:tt)* ) => { ... };
}

The pattern macro builds a crate::Pattern which is used by the crate::SearchTrie. The pattern macro takes the following parameters:

Regex(std::string::String) (required) - The regular expression to use for matching (note this is not a full regular expression. Only some parts of regular expressions are implemented like wildcards and carets). See crate::Pattern for full details.

PatternId(usize) (required) - The id for this pattern. This id is returned in the crate::Match array if this match occurs when the crate::SearchTrie::tmatch function is called.

IsMultiLine(bool) (optional) - If true is specified this pattern is a multi-line pattern meaning that wildcards can cross newlines. Otherwise newlines are not allowed in wildcard matches. If not specified this is true.

IsTerminationPattern(bool) (optional) - If true, this is a termination pattern meaning that if it is found, when the crate::SearchTrie::tmatch function is called, matching will terminate and the matches found up to that point in the text will be returned. If not specified this is false.

IsCaseSensitive(bool) (optional) - If true only case sensitive matches are returned for this pattern. Otherwise, case-insensitive matches are also returned. If not specified, this is false.

Return

Returns Ok(Pattern) on success and on error a [bmw_err::Error] is returned.

Errors

[bmw_err::ErrKind::Configuration] - If any other options are specified or if there are duplicates.

[bmw_err::ErrKind::Configuration] - If a Regex or PatternId is not specified.

[bmw_err::ErrKind::Configuration] - If both IsCaseSensitive and IsTerminationPattern are specified. (not currently supported)

Examples

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         // build a suffix tree with three patterns
         let mut search_trie = search_trie!(
                 vec![
                         // create a pattern that's multi-line and case sensitive
                         pattern!(
                             Regex("p1".to_string()),
                             PatternId(0),
                             IsTerminationPattern(false),
                             IsMultiLine(true),
                             IsCaseSensitive(true)
                         )?,
                         // create a pattern that is a termination pattern
                         pattern!(
                             Regex("p2".to_string()),
                             PatternId(1),
                             IsTerminationPattern(true),
                             IsCaseSensitive(false)
                         )?,
                         // use defaults (case insensitive, not termination and multiline true)
                         pattern!(Regex("p3".to_string()), PatternId(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildCardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [tmatch!()?; 10];

         // run the match for the input text b"p1p2".
         let count = search_trie.tmatch(b"p1p2", &mut matches)?;

         // assert that two matches were returned "p1" and "p2"
         // and that their start/end/id is correct.
         info!("count={}", count)?;
         assert_eq!(count, 2);
         assert_eq!(matches[1].id(), 0);
         assert_eq!(matches[1].start(), 0);
         assert_eq!(matches[1].end(), 2);
         assert_eq!(matches[0].id(), 1);
         assert_eq!(matches[0].start(), 2);
         assert_eq!(matches[0].end(), 4);

         Ok(())
 }