Macro bmw_util::queue_sync

macro_rules! queue_sync {
    ( $size:expr, $default:expr ) => { ... };
}

The crate::queue_sync macro is the sync version of crate::queue. This macro creates a crate::Queue with the specified parameters. The queue is returned as a impl Queue<T> + Send + Sync.

Input Parameters

size (usize) (required) - the size of the underlying array

default ([bmw_ser::Serializable]) (required) - a reference to the value to initialize the array with for the queue, these values are never used, but a default is needed to initialize the underlying array.

Return

Returns Ok(impl Queue<T> + Send + Sync) on success and a [bmw_err::Error] on failure.

Errors

[bmw_err::ErrKind::IllegalArgument] - if the size is 0.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::*;

 fn main() -> Result<(), Error> {
         let mut queue = queue_sync!(10, &0)?;

         for i in 0..10 {
                 queue.enqueue(i)?;
         }

         for i in 0..10 {
                 let v = queue.dequeue().unwrap();
                 assert_eq!(v, &i);
         }

         Ok(())
 }