Macro bmw_util::search_trie

macro_rules! search_trie {
    ( $patterns:expr, $($config:tt)*) => { ... };
}

The search_trie macro builds a crate::SearchTrie which can be used to match multiple patterns for a given text in a performant way.

Input Parameters

List<Pattern> (required) - The list of crate::Patterns that this crate::SearchTrie will use to match.

TerminationLength (usize) (optional) - The length in bytes at which matching will terminate.

MaxWildCardLength (usize) (optional) - The maximum length in bytes of a wild card match.

Return

Returns Ok(impl SearchTrie + Send + Sync) on success and on error a [bmw_err::Error] is returned.

Errors

[bmw_err::ErrKind::IllegalArgument] - If one of the regular expressions is invalid. or the length of the patterns list is 0.

Examples

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         // build a suffix tree with three patterns
         let mut search_trie = search_trie!(
                 vec![
                         pattern!(Regex("p1".to_string()), PatternId(0))?,
                         pattern!(Regex("p2".to_string()), PatternId(1))?,
                         pattern!(Regex("p3".to_string()), PatternId(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildCardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [tmatch!()?; 10];

         // run the match for the input text b"p1p2".
         let count = search_trie.tmatch(b"p1p2", &mut matches)?;

         // assert that two matches were returned "p1" and "p2"
         // and that their start/end/id is correct.
         info!("count={}", count)?;
         assert_eq!(count, 2);
         assert_eq!(matches[0].id(), 0);
         assert_eq!(matches[0].start(), 0);
         assert_eq!(matches[0].end(), 2);
         assert_eq!(matches[1].id(), 1);
         assert_eq!(matches[1].start(), 2);
         assert_eq!(matches[1].end(), 4);

         Ok(())
 }

Wild card match

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         // build a suffix tree with a wild card
         let mut search_trie = search_trie!(
                 vec![
                         pattern!(Regex("p1".to_string()), PatternId(0))?,
                         pattern!(Regex("p2.*test".to_string()), PatternId(1))?,
                         pattern!(Regex("p3".to_string()), PatternId(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildCardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [UtilBuilder::build_match(vec![])?; 10];

         // run the match for the input text b"p1p2". Only "p1" matches this time.
         let count = search_trie.tmatch(b"p1p2", &mut matches)?;
         assert_eq!(count, 1);

         // run the match for the input text b"p1p2xxxxxxtest1". Now the wildcard
         // match succeeds to two matches are returned.
         let count = search_trie.tmatch(b"p1p2xxxxxxtest1", &mut matches)?;
         assert_eq!(count, 2);

         Ok(())
 }

Single character wild card

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         // build a suffix tree with a wild card
         let mut search_trie = search_trie!(
                 vec![
                         pattern!(Regex("p1".to_string()), PatternId(0))?,
                         pattern!(Regex("p2.test".to_string()), PatternId(1))?,
                         pattern!(Regex("p3".to_string()), PatternId(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildCardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [tmatch!()?; 10];

         // run the match for the input text b"p1p2". Only "p1" matches this time.
         let count = search_trie.tmatch(b"p1p2", &mut matches)?;
         assert_eq!(count, 1);

         // run the match for the input text b"p1p2xxxxxxtest1". Now the wildcard
         // match doesn't succeed because it's a single char match. One match is returned.
         let count = search_trie.tmatch(b"p1p2xxxxxxtest1", &mut matches)?;
         assert_eq!(count, 1);

         // run it with a single char and see that it matches pattern two.
         let count = search_trie.tmatch(b"p1p2xtestx", &mut matches)?;
         assert_eq!(count, 2);

         Ok(())
 }

Match at the beginning of the text

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {      
         // build a suffix tree with a wild card
         let mut search_trie = search_trie!(
                 vec![
                         pattern!(Regex("p1".to_string()), PatternId(0))?,
                         pattern!(Regex("^p2".to_string()), PatternId(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildCardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [tmatch!()?; 10];

         // run the match for the input text b"p1p2". Only "p1" matches this time
         // because p2 is not at the start
         let count = search_trie.tmatch(b"p1p2", &mut matches)?;
         assert_eq!(count, 1);

         // since p2 is at the beginning, both match
         let count = search_trie.tmatch(b"p2p1", &mut matches)?;
         assert_eq!(count, 2);

         Ok(())
 }