Macro bmw_util::slab_allocator

macro_rules! slab_allocator {
    () => { ... };
    ($slab_count:expr) => { ... };
    ($slab_count:expr, $slab_size:expr) => { ... };
}

The crate::slab_allocator macro allows for creation of a slab allocator that may be used by the data structures in this crate. If no parameters are specified, the default values of slab_size equal to 1_024 and slab_count of 10_240 are used. If one paramater is specifed, it is used as slab_count and the default is used for slab_size. If two parameters are specified, it is used as slab_count and slab_size respectively. The slab_count is the total number of slabs in this slab allocator. It is important to note that additional slabs may not be added after startup. The slab_size is the size in bytes of the slabs in this instance.

Errors

crate::slab_allocator returns a crate::SlabAllocator on success. On error, a bmw_err::Error is returned.

bmw_err::ErrorKind::IllegalArgument if the slab_size is less than 48 bytes or if slab_count is equal to 0.

bmw_err::ErrorKind::IllegalState if the thread local global slab allocator for this thread has already been initialized.

Examples

 use bmw_err::*;
 use bmw_util::{slab_allocator, hashtable};

 fn main() -> Result<(), Error> {
     let slab_allocator1 = slab_allocator!()?;
     let slab_allocator2 = slab_allocator!(8_000)?;
     let slab_allocator3 = slab_allocator!(3_000, 2_048)?;
     let mut hashtable1 = hashtable!(1_000, 0.9, slab_allocator1)?;
     let mut hashtable2 = hashtable!(5_000, 0.8, slab_allocator2)?;
     let mut hashtable3 = hashtable!(3_000, 0.7, slab_allocator3)?;

     hashtable1.insert(&1, &2)?;
     hashtable2.insert(&10, &20)?;
     hashtable3.insert(&100, &200)?;

     // ...
     Ok(())
 }

Note: the slab allocator returned by this macro is distinct from the global thread local slab allocator that is initialized by the crate::init_slab_allocator macro.