Macro bmw_util::suffix_tree

macro_rules! suffix_tree {
    ( $patterns:expr, $( $suffix_items:expr ),* ) => { ... };
}

The suffix_tree macro builds a crate::SuffixTree which can be used to match multiple patterns for a given text in a performant way. The suffix_tree macro takes the following parameters:

List<Pattern> (required) - The list of crate::Patterns that this crate::SuffixTree will use to match.

TerminationLength(usize) (optional) - The length in bytes at which matching will terminate.

MaxWildcardLength(usize) (optional) - The maximum length in bytes of a wild card match.

Return

Returns Ok(SuffixTre) on success and on error a bmw_err::Error is returned.

Errors

bmw_err::ErrorKind::IllegalArgument - If one of the regular expressions is invalid. or the length of the patterns list is 0.

Examples

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         // build a suffix tree with three patterns
         let mut suffix_tree = suffix_tree!(
                 list![
                         pattern!(Regex("p1"), Id(0))?,
                         pattern!(Regex("p2"), Id(1))?,
                         pattern!(Regex("p3"), Id(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildcardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [Builder::build_match_default(); 10];

         // run the match for the input text b"p1p2".
         let count = suffix_tree.tmatch(b"p1p2", &mut matches)?;

         // assert that two matches were returned "p1" and "p2"
         // and that their start/end/id is correct.
         info!("count={}", count)?;
         assert_eq!(count, 2);
         assert_eq!(matches[0].id(), 0);
         assert_eq!(matches[0].start(), 0);
         assert_eq!(matches[0].end(), 2);
         assert_eq!(matches[1].id(), 1);
         assert_eq!(matches[1].start(), 2);
         assert_eq!(matches[1].end(), 4);

         Ok(())
 }

Wild card match

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         // build a suffix tree with a wild card
         let mut suffix_tree = suffix_tree!(
                 list![
                         pattern!(Regex("p1"), Id(0))?,
                         pattern!(Regex("p2.*test"), Id(1))?,
                         pattern!(Regex("p3"), Id(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildcardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [Builder::build_match_default(); 10];

         // run the match for the input text b"p1p2". Only "p1" matches this time.
         let count = suffix_tree.tmatch(b"p1p2", &mut matches)?;
         assert_eq!(count, 1);

         // run the match for the input text b"p1p2xxxxxxtest1". Now the wildcard
         // match succeeds to two matches are returned.
         let count = suffix_tree.tmatch(b"p1p2xxxxxxtest1", &mut matches)?;
         assert_eq!(count, 2);

         Ok(())
 }

Single character wild card

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         // build a suffix tree with a wild card
         let mut suffix_tree = suffix_tree!(
                 list![
                         pattern!(Regex("p1"), Id(0))?,
                         pattern!(Regex("p2.test"), Id(1))?,
                         pattern!(Regex("p3"), Id(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildcardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [Builder::build_match_default(); 10];

         // run the match for the input text b"p1p2". Only "p1" matches this time.
         let count = suffix_tree.tmatch(b"p1p2", &mut matches)?;
         assert_eq!(count, 1);

         // run the match for the input text b"p1p2xxxxxxtest1". Now the wildcard
         // match doesn't succeed because it's a single char match. One match is returned.
         let count = suffix_tree.tmatch(b"p1p2xxxxxxtest1", &mut matches)?;
         assert_eq!(count, 1);

         // run it with a single char and see that it matches pattern two.
         let count = suffix_tree.tmatch(b"p1p2xtestx", &mut matches)?;
         assert_eq!(count, 2);

         Ok(())
 }

Match at the beginning of the text

 use bmw_util::*;
 use bmw_err::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {      
         // build a suffix tree with a wild card
         let mut suffix_tree = suffix_tree!(
                 list![
                         pattern!(Regex("p1"), Id(0))?,
                         pattern!(Regex("^p2"), Id(2))?
                 ],
                 TerminationLength(1_000),
                 MaxWildcardLength(100)
         )?;

         // create a matches array for the suffix tree to return matches in
         let mut matches = [Builder::build_match_default(); 10];

         // run the match for the input text b"p1p2". Only "p1" matches this time
         // because p2 is not at the start
         let count = suffix_tree.tmatch(b"p1p2", &mut matches)?;
         assert_eq!(count, 1);

         // since p2 is at the beginning, both match
         let count = suffix_tree.tmatch(b"p2p1", &mut matches)?;
         assert_eq!(count, 2);

         Ok(())
 }