Macro bmw_util::thread_pool

macro_rules! thread_pool {
    ( $( $match_items:tt)* ) => { ... };
}

Macro used to configure/build a thread pool. Threadpools can be used to execute tasks in a set of threads that is configurable.

Input Parameters

MaxSize(usize) (optional) - the maximum size that the thread pool, in terms of number of threads, may grow to. The default value is MinSize.

MinSize(usize) (optional) - the minumum size, in threads, that the thread pool will maintain. The thread pool will add threads up until MaxSize, but never go below MinSize threads. The default value is 1.

SyncChannelSize(usize) (optional) - the size of the internal sync_channel used to send tasks to the thread pool threads for execution. The default is 10.

Return value

Upon success, this macro will return a crate::ThreadPoolHandle.

Errors

[bmw_err::ErrKind::Configuration] - If the configuration contained parameters other than MaxSize, MinSize, or SyncChannelSize.

[bmw_err::ErrKind::Configuration] - If the configuration contained duplicate parameters.

[bmw_err::ErrKind::Configuration] - If the configuration of MinSize is 0 or if MaxSize is less than MinSize.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_test::*;
 use bmw_util::*;

 info!();

 fn main() -> Result<(), Error> {
     info!("testing thread_pool macro")?;

     // if only MinSize is specified, the thread pool maintains
     // MinSize threads (in this case 4) at all times.
     let mut tp = thread_pool!(MinSize(4))?;
     tp.set_on_panic(move |id, e| -> Result<(), Error> {
         info!("PANIC: id={},e={:?}", id, e)?;
         Ok(())
     })?;

     tp.start()?;

     execute!(tp, {
          info!("executing a thread")?;
          Ok(())
     })?;

     Ok(())
 }

Here’s an example with other values configured. This also demonstrates how to block on a task in another thread.

 use bmw_err::*;
 use bmw_log::*;
 use bmw_test::*;
 use bmw_util::*;

 info!();

 fn main() -> Result<(), Error> {
     info!("testing thread_pool macro")?;

     let mut tp = thread_pool!(MinSize(4), MaxSize(5), SyncChannelSize(15))?;
     tp.set_on_panic(move |id, e| -> Result<(), Error> {
         info!("PANIC: id={},e={:?}", id, e)?;
         Ok(())
     })?;
     
     tp.start()?;
     
     let mut tph = execute!(tp, {
          info!("executing a thread")?;
          Ok(1)
     })?;

     let result = tph.block_on();

     info!("Thread {} resulted in a value of {:?}", tph.id(), result)?;

     assert_eq!(result, PoolResult::Ok(1));
         
     Ok(())
 }

Here’s an example handling a thread panic.

 use bmw_err::*;
 use bmw_log::*;
 use bmw_test::*;
 use bmw_util::*;
 use bmw_deps::rand::random;
 use std::sync::mpsc::sync_channel;

 info!();

 fn main() -> Result<(), Error> {
     info!("testing thread_pool macro with panic")?;
     let mut id_box = lock_box!(random())?;
     let id_box_clone = id_box.clone();
     let (tx, rx) = sync_channel(1);

     let mut tp = thread_pool!(MinSize(4), MaxSize(5), SyncChannelSize(15))?;
     tp.set_on_panic(move |id, e| -> Result<(), Error> {
         info!("PANIC: id={},e={:?}", id, e)?;

         wlock!(id_box) = id;
         tx.send(());

         Ok(())
     })?;
  
     tp.start()?;

     let mut tph = execute!(tp, {
          info!("executing a thread")?;
          if true {
                 panic!("12345");
          }
          Ok(1)
     })?;
     
     let result = tph.block_on();
     
     info!("Thread {} resulted in a value of {:?}", tph.id(), result)?;
     
     assert_eq!(
         result,
         PoolResult::Err(
             err!(
                 ErrKind::ThreadPanic,
                 "thread pool panic: receiving on a closed channel"
             )
         )
     );
     rx.recv()?;
     assert_eq!(rlock!(id_box_clone), tph.id());
         
     Ok(())
 }