Struct bmw_util::LogConfig

pub struct LogConfig {

    pub colors: LogConfigOption,
    pub stdout: LogConfigOption,
    pub max_size_bytes: LogConfigOption,
    pub max_age_millis: LogConfigOption,
    pub timestamp: LogConfigOption,
    pub level: LogConfigOption,
    pub line_num: LogConfigOption,
    pub show_millis: LogConfigOption,
    pub auto_rotate: LogConfigOption,
    pub file_path: LogConfigOption,
    pub show_bt: LogConfigOption,
    pub line_num_data_max_len: LogConfigOption,
    pub delete_rotation: LogConfigOption,
    pub file_header: LogConfigOption,
    pub debug_invalid_metadata: bool,
    pub debug_invalid_os_str: bool,
    pub debug_lineno_none: bool,

}

The log configuration struct. Logs can only be built through the crate::LogBuilder::build function. This is the only parameter to that function. An example configuration with all parameters explicitly specified might look like this:

 use bmw_log::LogConfigOption::*;
 use bmw_log::LogConfig;

 let config = LogConfig {
     colors: Colors(true),
     stdout: Stdout(true),
     max_size_bytes: MaxSizeBytes(1024 * 1024 * 5),
     max_age_millis: MaxAgeMillis(1000 * 30 * 60),
     timestamp: Timestamp(true),
     level: Level(true),
     line_num: LineNum(false),
     show_millis: ShowMillis(false),
     auto_rotate: AutoRotate(true),
     file_path: FilePath(None),
     show_bt: ShowBt(true),
     line_num_data_max_len: LineNumDataMaxLen(20),
     delete_rotation: DeleteRotation(false),
     file_header: FileHeader("BMW Log V1.1".to_string()),
     debug_invalid_metadata: false,
     debug_invalid_os_str: false,
     debug_lineno_none: false,
 };

Generally speaking the configurations are specified using the core::default::Default trait which is implemented for LogConfig. An example might look like this:

 use bmw_log::LogConfigOption::*;
 use bmw_log::LogConfig;
 use std::path::PathBuf;

 let config = LogConfig {
     colors: Colors(false),
     stdout: Stdout(false),
     line_num: LineNum(false),
     file_path: FilePath(Some(PathBuf::from("/path/to/my/log.log".to_string()))),
     ..Default::default()
 };

Fields
colors: LogConfigOption

See LogConfigOption::Colors. The default value is Colors(true).

stdout: LogConfigOption

See LogConfigOption::Stdout. The default value is Stdout(true).

max_size_bytes: LogConfigOption

See LogConfigOption::MaxSizeBytes. The default value is MaxSizeBytes(1024 * 1024) or 1 mb.

max_age_millis: LogConfigOption

See LogConfigOption::MaxAgeMillis. The default value is MaxAgeMillis(60 * 60 * 1000) or 1 hour.

timestamp: LogConfigOption

See LogConfigOption::Timestamp. The default value is Timestamp(true).

level: LogConfigOption

See LogConfigOption::Level. The default value is Level(true).

line_num: LogConfigOption

See LogConfigOption::LineNum. The default value is LineNum(true).

show_millis: LogConfigOption

See LogConfigOption::ShowMillis. The default value is ShowMillis(true).

auto_rotate: LogConfigOption

See LogConfigOption::AutoRotate. The default value is AutoRotate(true).

file_path: LogConfigOption

See LogConfigOption::FilePath. The default value is FilePath(None).

show_bt: LogConfigOption

See LogConfigOption::ShowBt. The default value is ShowBt(true).

line_num_data_max_len: LogConfigOption

See LogConfigOption::LineNumDataMaxLen. The default value is LinNumDataMaxLen(25) or 25 bytes.

delete_rotation: LogConfigOption

See LogConfigOption::DeleteRotation. The default value is DeleteRotation(false).

file_header: LogConfigOption

See LogConfigOption::FileHeader. The default value is FileHeader(“”.to_string()) or no file header.

debug_invalid_metadata: bool

Parameter used in testing. Must always be set false in normal use.

debug_invalid_os_str: bool

Parameter used in testing. Must always be set false in normal use.

debug_lineno_none: bool

Parameter used in testing. Must always be set false in normal use.