Trait bmw_util::Hashset

pub trait Hashset<K>: Debug + DynClone
where
    K: Serializable + Clone,
{
    // Required methods
    fn max_load_factor(&self) -> f64;
    fn max_entries(&self) -> usize;
    fn insert(&mut self, key: &K) -> Result<(), Error>;
    fn contains(&self, key: &K) -> Result<bool, Error>;
    fn remove(&mut self, key: &K) -> Result<bool, Error>;
    fn size(&self) -> usize;
    fn clear(&mut self) -> Result<(), Error>;
    fn iter<'a>(&'a self) -> HashsetIterator<'a, K> ;
    fn slabs(
        &self,
    ) -> Result<Option<Box<dyn LockBox<Box<dyn SlabAllocator + Send + Sync>>>>, Error>;
}

The hashset trait. See crate::hashset for working examples.

Required Methods

fn max_load_factor(&self) -> f64

Returns the maximum load factor as configured for this crate::Hashset.

fn max_entries(&self) -> usize

Returns the maximum entries as configured for this crate::Hashset.

fn insert(&mut self, key: &K) -> Result<(), Error>

Insert a key into this hashset.

fn contains(&self, key: &K) -> Result<bool, Error>

If key is present this function returns true, otherwise false.

fn remove(&mut self, key: &K) -> Result<bool, Error>

Remove the specified key from this hashset.

fn size(&self) -> usize

Returns the size of this hashset.

fn clear(&mut self) -> Result<(), Error>

Clear all items, reinitialized the entry array, and free the slabs associated with this hashset.

fn iter<'a>(&'a self) -> HashsetIterator<'a, K>

Returns an std::iter::Iterator to iterate through this hashset.

fn slabs( &self, ) -> Result<Option<Box<dyn LockBox<Box<dyn SlabAllocator + Send + Sync>>>>, Error>