Trait bmw_util::Hashtable

pub trait Hashtable<K, V>: Debug + DynClone
where
    K: Serializable + Clone,
    V: Serializable,
{

    // Required methods
    fn max_load_factor(&self) -> f64;
    fn max_entries(&self) -> usize;
    fn insert(&mut self, key: &K, value: &V) -> Result<(), Error>;
    fn get(&self, key: &K) -> Result<Option<V>, Error>;
    fn remove(&mut self, key: &K) -> Result<Option<V>, Error>;
    fn size(&self) -> usize;
    fn clear(&mut self) -> Result<(), Error>;
    fn iter<'a>(&'a self) -> HashtableIterator<'a, K, V> ;
    fn bring_to_front(&mut self, key: &K) -> Result<(), Error>;
    fn remove_oldest(&mut self) -> Result<(), Error>;
    fn raw_read(
        &self,
        key: &K,
        offset: usize,
        data: &mut [u8; 412],
    ) -> Result<bool, Error>;
    fn raw_write(
        &mut self,
        key: &K,
        offset: usize,
        data: &[u8; 412],
        len: usize,
    ) -> Result<(), Error>;
    fn slabs(
        &self,
    ) -> Result<Option<Box<dyn LockBox<Box<dyn SlabAllocator + Send + Sync>>>>, Error>;

}

The hashtable trait. See crate::hashtable for working examples.

Required Methods

fn max_load_factor(&self) -> f64

Returns the maximum load factor as configured for this crate::Hashtable.

fn max_entries(&self) -> usize

Returns the maximum entries as configured for this crate::Hashtable.

fn insert(&mut self, key: &K, value: &V) -> Result<(), Error>

Insert a key/value pair into the hashtable.

fn get(&self, key: &K) -> Result<Option<V>, Error>

Get the value associated with the specified key.

fn remove(&mut self, key: &K) -> Result<Option<V>, Error>

Remove the specified key from the hashtable.

fn size(&self) -> usize

Return the size of the hashtable.

fn clear(&mut self) -> Result<(), Error>

Clear all items, reinitialized the entry array, and free the slabs associated with this hashtable.

fn iter<'a>(&'a self) -> HashtableIterator<'a, K, V>

Returns an std::iter::Iterator to iterate through this hashtable.

fn bring_to_front(&mut self, key: &K) -> Result<(), Error>

Bring the entry to the front of the list for deletion purposes in a cache.

fn remove_oldest(&mut self) -> Result<(), Error>

Remove the oldest entry in the hashtable.

fn raw_read( &self, key: &K, offset: usize, data: &mut [u8; 412], ) -> Result<bool, Error>

Get raw data and store it in data with given offset.

fn raw_write( &mut self, key: &K, offset: usize, data: &[u8; 412], len: usize, ) -> Result<(), Error>

Write raw data from data with given offset.

fn slabs( &self, ) -> Result<Option<Box<dyn LockBox<Box<dyn SlabAllocator + Send + Sync>>>>, Error>

Gets the slab allocator associated with this Hashtable or None if the global slab allocator is used.