Trait bmw_util::List

pub trait List<V>: DynClone + Debug {
    // Required methods
    fn push(&mut self, value: V) -> Result<(), Error>;
    fn iter<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a>
       where V: Serializable + Clone;
    fn iter_rev<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a>
       where V: Serializable + Clone;
    fn delete_head(&mut self) -> Result<(), Error>;
    fn size(&self) -> usize;
    fn clear(&mut self) -> Result<(), Error>;
}

The list trait is implemented by both crate::list and crate::array_list. See these macros for detailed examples.

Required Methods

fn push(&mut self, value: V) -> Result<(), Error>

push a value onto the end of the list

fn iter<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a>
where V: Serializable + Clone,

return an iterator that can iterate through the list

fn iter_rev<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a>
where V: Serializable + Clone,

return an iterator that can iterate through the list in reverse order

fn delete_head(&mut self) -> Result<(), Error>

delete the head of the list

fn size(&self) -> usize

return the size of the list

fn clear(&mut self) -> Result<(), Error>

clear all items from the list