Trait bmw_util::Lock

pub trait Lock<T>: Send + Sync + Debugwhere
    T: Send + Sync,{
    // Required methods
    fn wlock(&mut self) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>;
    fn rlock(&self) -> Result<RwLockReadGuardWrapper<'_, T>, Error>;
    fn clone(&self) -> Self;
}

Wrapper around the lock functionalities used by bmw in std::sync rust libraries. The main benefits are the simplified interface and the fact that if a thread attempts to obtain a lock twice, an error will be thrown instead of a thread panic. This is implemented through a thread local Hashset which keeps track of the guards used by the lock removes an entry for them when the guard is dropped.

Examples

 use bmw_util::*;
 use bmw_err::*;
 use std::time::Duration;
 use std::thread::{sleep,spawn};

 fn test() -> Result<(), Error> {
     let mut lock = lock!(1)?;
     let lock_clone = lock.clone();

     spawn(move || -> Result<(), Error> {
         let mut x = lock.wlock()?;
         assert_eq!(**(x.guard()), 1);
         sleep(Duration::from_millis(3000));
         **(x.guard()) = 2;
         Ok(())
     });

     sleep(Duration::from_millis(1000));
     let x = lock_clone.rlock()?;
     assert_eq!(**(x.guard()), 2);
     Ok(())
 }

Required Methods

fn wlock(&mut self) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>

obtain a write lock and corresponding std::sync::RwLockWriteGuard for this crate::Lock.

fn rlock(&self) -> Result<RwLockReadGuardWrapper<'_, T>, Error>

obtain a read lock and corresponding std::sync::RwLockReadGuard for this crate::Lock.

fn clone(&self) -> Self

Clone this crate::Lock.