Trait bmw_util::LockBox

pub trait LockBox<T>: Send + Sync + Debug
where
    T: Send + Sync,
{
    // Required methods
    fn wlock(&mut self) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>;
    fn rlock(&self) -> Result<RwLockReadGuardWrapper<'_, T>, Error>;
    fn wlock_ignore_poison(
        &mut self,
    ) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>;
    fn rlock_ignore_poison(
        &self,
    ) -> Result<RwLockReadGuardWrapper<'_, T>, Error>;
    fn danger_to_usize(&self) -> usize;
    fn inner(&self) -> Arc<RwLock<T>>;
    fn id(&self) -> u128;
}

A Lock which can be used to send data to and from threads which is also stored in a box so it can be stored in structures as well. See crate::lock_box.

Required Methods

fn wlock(&mut self) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>

obtain a write lock and corresponding std::sync::RwLockWriteGuard for this crate::LockBox.

fn rlock(&self) -> Result<RwLockReadGuardWrapper<'_, T>, Error>

obtain a read lock and corresponding std::sync::RwLockReadGuard for this crate::LockBox.

fn wlock_ignore_poison( &mut self, ) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>

Same as crate::LockBox::wlock except that any poison errors are ignored by calling the underlying into_inner() fn.

fn rlock_ignore_poison(&self) -> Result<RwLockReadGuardWrapper<'_, T>, Error>

Same as crate::LockBox::rlock except that any poison errors are ignored by calling the underlying into_inner() fn.

fn danger_to_usize(&self) -> usize

consume the inner Arc and return a usize value. This function is dangerous because it potentially leaks memory. The usize must be rebuilt into a lockbox that can then be dropped via the crate::lock_box_from_usize function.

fn inner(&self) -> Arc<RwLock<T>>

return the inner data holder.

fn id(&self) -> u128

return the id for this lockbox.