Trait bmw_util::RawHashtableIterator

pub trait RawHashtableIterator<Item = (Vec<u8>, Vec<u8>)>: Iterator { }

An iterator for iterating through raw data in this crate::StaticHashtable. This is distinct from the iterator that is implemented as crate::StaticHashtable’s std::iter::IntoIterator trait which returns the serialized and not raw data.

Examples

 use bmw_err::*;
 use bmw_util::hashtable;

 fn test() -> Result<(), Error> {
     // create a hashtable
     let mut hashtable = hashtable!()?;
     // need to do a empty insert so that the types can be inferred. From here on we use raw
     // operations
     assert!(hashtable.insert(&(), &()).is_err());
     let key = [0u8, 1u8, 123u8];
     let value = [10u8];
     let hash = 123usize;
     hashtable.insert_raw(&key, hash, &value)?;

     let key = [1u8, 1u8, 125u8];
     let value = [14u8];
     let hash = 125usize;
     hashtable.insert_raw(&key, hash, &value)?;

     let mut count = 0;
     for (k,v) in hashtable.iter_raw() {
         if v == [14u8] {
             assert_eq!(k, [1u8, 1u8, 125u8]);
             count += 1;
         } else if v == [10u8] {
             assert_eq!(k, [0u8, 1u8, 123u8]);
             count += 1;
         }
     }

     assert_eq!(count, 2);

     Ok(())
 }