Trait bmw_util::SearchTrie

pub trait SearchTrie: DynClone {
    // Required method
    fn tmatch(
        &mut self,
        text: &[u8],
        matches: &mut [Match],
    ) -> Result<usize, Error>;
}

The search trie is a Trie that can be used to efficiently search for multiple matches within a single string. It is designed to be fast and not allocate additional memory after it has been initialized. It is important to note that this trie is not space efficient though. See crate::search_trie! which is the macro that builds this data structure.

Required Methods

fn tmatch(&mut self, text: &[u8], matches: &mut [Match]) -> Result<usize, Error>

return matches associated with the supplied text for this crate::SearchTrie. Matches are returned in the matches array supplied by the caller. The result is the number of matches found or a [bmw_err::Error] if an error occurs.