Trait bmw_util::Serializable

pub trait Serializable {
    fn read<R>(reader: &mut R) -> Result<Self, Error>
    where
        R: Reader;
    fn write<W>(&self, writer: &mut W) -> Result<(), Error>
    where
        W: Writer;
}

This is the trait used by all data structures to serialize and deserialize data. Anything stored in them must implement this trait. Commonly needed implementations are built in the ser module in this crate. These include Vec, String, integer types among other things.

Required Methods

fn read<R>(reader: &mut R) -> Result<Self, Error> where
R: Reader,

read data from the reader and build the underlying type represented by that data.

fn write<W>(&self, writer: &mut W) -> Result<(), Error> where
W: Writer,

write data to the writer representing the underlying type.