Trait bmw_util::SlabAllocator

pub trait SlabAllocator: DynClone + Debug {
    // Required methods
    fn is_init(&self) -> bool;
    fn allocate<'a>(&'a mut self) -> Result<SlabMut<'a>, Error>;
    fn free(&mut self, id: usize) -> Result<(), Error>;
    fn get<'a>(&'a self, id: usize) -> Result<Slab<'a>, Error>;
    fn get_mut<'a>(&'a mut self, id: usize) -> Result<SlabMut<'a>, Error>;
    fn free_count(&self) -> Result<usize, Error>;
    fn slab_size(&self) -> Result<usize, Error>;
    fn slab_count(&self) -> Result<usize, Error>;
    fn init(&mut self, config: SlabAllocatorConfig) -> Result<(), Error>;
}

The slab allocator trait defines the operations available for a Slab Allocator. See crate::slab_allocator for working examples.

Required Methods

fn is_init(&self) -> bool

If the slab allocator has been initialized, return true, otherwise, false.

fn allocate<'a>(&'a mut self) -> Result<SlabMut<'a>, Error>

Allocate a slab or return an error if one is not available.

fn free(&mut self, id: usize) -> Result<(), Error>

Free a slab. Return an error if the slab cannot be freed.

fn get<'a>(&'a self, id: usize) -> Result<Slab<'a>, Error>

Get an immutible reference to a slab.

fn get_mut<'a>(&'a mut self, id: usize) -> Result<SlabMut<'a>, Error>

Get a mutable reference to a slab.

fn free_count(&self) -> Result<usize, Error>

Returns the number of free slabs this crate::SlabAllocator has remaining.

fn slab_size(&self) -> Result<usize, Error>

Returns the configured slab_size for this crate::SlabAllocator.

fn slab_count(&self) -> Result<usize, Error>

Returns the configured slab_count for this crate::SlabAllocator.

fn init(&mut self, config: SlabAllocatorConfig) -> Result<(), Error>

Initializes the crate::SlabAllocator with the given config. See crate::SlabAllocatorConfig for further details.