Trait bmw_util::SortableList

pub trait SortableList<V>: List<V> + DynClone {
    // Required methods
    fn sort(&mut self) -> Result<(), Error>
       where V: Ord;
    fn sort_unstable(&mut self) -> Result<(), Error>
       where V: Ord;
}

The sortable list allows for efficient sorting of both the crate::array_list and crate::list.

Required Methods

fn sort(&mut self) -> Result<(), Error>
where V: Ord,

sort with a stable sorting algorithm

fn sort_unstable(&mut self) -> Result<(), Error>
where V: Ord,

sort with an unstable sorting algorithm. unstable sort is significantly faster and should be used when stable sorting (ording of equal values consistent) is not required.