Trait bmw_util::StaticHashtable

pub trait StaticHashtable<K, V> where
    K: Serializable + Hash,
    V: Serializable, {

    fn config(&self) -> StaticHashtableConfig;
    fn insert(
        &mut self, 
        ctx: &mut Context, 
        key: &K, 
        value: &V
    ) -> Result<(), Error>;
    fn get(&self, ctx: &mut Context, key: &K) -> Result<Option<V>, Error>;
    fn remove(&mut self, ctx: &mut Context, key: &K) -> Result<Option<V>, Error>;
    fn get_raw<'b>(
        &'b self, 
        ctx: &mut Context, 
        key: &[u8], 
        hash: usize
    ) -> Result<Option<Slab<'b>>, Error>;
    fn insert_raw(
        &mut self, 
        ctx: &mut Context, 
        key: &[u8], 
        hash: usize, 
        value: &[u8]
    ) -> Result<(), Error>;
    fn remove_raw(
        &mut self, 
        ctx: &mut Context, 
        key: &[u8], 
        hash: usize
    ) -> Result<bool, Error>;
    fn iter_raw<'b>(&'b self, ctx: &mut Context) -> RawHashtableIterator<'b>;
    fn size(&self, ctx: &mut Context) -> usize;
    fn first_entry(&self, ctx: &mut Context) -> usize;
    fn slab<'b>(
        &'b self, 
        ctx: &mut Context, 
        id: usize
    ) -> Result<Slab<'b>, Error>;
    fn read_kv(
        &self, 
        ctx: &mut Context, 
        slab_id: usize
    ) -> Result<(K, V), Error>;
    fn get_array(&self, ctx: &mut Context) -> &Vec<usize>;
    fn clear(&mut self, ctx: &mut Context) -> Result<(), Error>;

}

The crate::StaticHashtable trait defines the public interface to the static hashtable. The hashtable in this crate uses linear probing to handle collisions and it cannot be resized after it is intialized. Configuration of the hashtable is done via the crate::StaticHashtableConfig struct. The shared implementation can be instantiated as a crate::StaticHashtable through the crate::StaticHashtableBuilder::build function or as a crate::StaticHashset through the crate::StaticHashsetBuilder::build function. Although there is a different interface for each, they are very similar and share most of the implementation code. In most cases, the data structures in this crate should be instantiated through the macros, but they can also be instantiated through the builder structs as well. See crate::hashtable.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::{ctx, hashtable};

 info!();

 fn main() -> Result<(), Error> {
     let ctx = ctx!();
     let mut hash = hashtable!()?;

     hash.insert(ctx, &1, &"abc".to_string())?;
     hash.insert(ctx, &3, &"def".to_string())?;

     for (k,v) in &hash {
         info!("k={},v={}", k, v)?;
     }

     Ok(())
 }

Required Methods

fn config(&self) -> StaticHashtableConfig

This function returns a copy of the underlying crate::StaticHashtableConfig struct.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::{ctx, hashtable};

 info!();

 fn main() -> Result<(), Error> {
     let ctx = ctx!();
     let mut hash = hashtable!(1_000, 0.9)?;

     hash.insert(ctx, &1, &"abc".to_string())?;
     let config = hash.config();
     assert_eq!(config.max_entries, 1_000);
     assert_eq!(config.max_load_factor, 0.9);

     Ok(())
 }

fn insert(&mut self, ctx: &mut Context, key: &K, value: &V) -> Result<(), Error>

This function inserts the key value pair specified. Both K/V must implement the crate::Serializable trait and key must implement the std::hash::Hash trait. The key and value are both copied into the table. Note that after this point, changes the underlying key or value will not be reflected in the stored value. This is in contrast to the Rust standard Hashtable which uses references.

Examples

 use bmw_err::*;
 use bmw_util::{ctx, hashtable};

 fn main() -> Result<(), Error> {
     let ctx = ctx!();
     let mut hash = hashtable!(1_000, 0.9)?;

     // since Vec implements Serializable, we can insert vec.
     hash.insert(ctx, &1, &vec![1u8,2u8,3u8,4u8])?;

     Ok(())
 }

fn get(&self, ctx: &mut Context, key: &K) -> Result<Option<V>, Error>

Get an element in the crate::StaticHashtable. A copy of the serialized value is returned or None if it is not found.

Examples

 use bmw_err::*;
 use bmw_util::{ctx, hashtable};

 fn main() -> Result<(), Error> {
     let ctx = ctx!(); // context
     let mut hash = hashtable!(1_000, 0.9)?;

     // since Vec implements Serializable, we can insert a Vec.
     hash.insert(ctx, &1, &vec![1u8,2u8,3u8,4u8])?;
     let v = hash.get(ctx, &1)?;
     assert_eq!(v.unwrap(), vec![1u8,2u8,3u8,4u8]);
     assert!(hash.get(ctx, &2)?.is_none());

     Ok(())
 }

fn remove(&mut self, ctx: &mut Context, key: &K) -> Result<Option<V>, Error>

Remove an element from the crate::StaticHashtable. Upon success, the element removed from the crate::StaticHashtable is returned. If the key is not found in the crate::StaticHashtable, None is returned. If an error occurs, bmw_err::Error is returned.

Examples

 use bmw_util::{ctx,hashtable};
 use bmw_err::Error;

 fn hashtable_remove() -> Result<(), Error> {
     let ctx = ctx!();
     let mut sh = hashtable!()?;
     assert_eq!(sh.get(ctx, &1)?, None);
     sh.insert(ctx, &1, &100)?;
     assert_eq!(sh.get(ctx, &1)?, Some(100));
     sh.remove(ctx, &1)?;
     assert_eq!(sh.get(ctx, &1)?, None);

     Ok(())
 }

fn get_raw<'b>(
&'b self,
ctx: &mut Context,
key: &[u8],
hash: usize
) -> Result<Option<Slab<'b>>, Error>

Get the raw crate::Slab data, if it exists, in this crate::StaticHashtable. On success return the slab, if the slab is not found, return None or if an error occurs, return bmw_err::Error.

Examples

 use bmw_util::{hashtable, ctx, hashtable_set_raw};
 use bmw_err::Error;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use bmw_log::usize;

 fn raw_ex() -> Result<(), Error> {
     let ctx = ctx!(); // create a context
     let mut sh = hashtable!()?; // create a hashtable
     hashtable_set_raw!(ctx, sh); // set raw mode

     // generate our own hash manually
     let mut hasher = DefaultHasher::new();
     (b"hi").hash(&mut hasher);
     let hash = hasher.finish();

     // insert
     sh.insert_raw(ctx, b"hi", usize!(hash), b"ok")?;

     // call get_raw to lookup our entry
     let slab = sh.get_raw(ctx, b"hi", usize!(hash))?.unwrap();

     // key = 104/105 (hi), value = 111/107 (ok)
     assert_eq!(
         slab.get()[0..36],
         [
             255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
             0, 0, 0, 0, 0, 0, 0, 2, 104, 105, 0, 0, 0, 0, 0, 0, 0, 2, 111, 107
         ]
     );

     Ok(())
 }

fn insert_raw(
&mut self,
ctx: &mut Context,
key: &[u8],
hash: usize,
value: &[u8]
) -> Result<(), Error>

Insert raw data into the crate::StaticHashtable. On success returns Ok(()), on failure returns bmw_err::Error.

Examples

 use bmw_util::{hashtable, ctx, hashtable_set_raw};
 use bmw_err::Error;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use bmw_log::usize;

 fn raw_ex() -> Result<(), Error> {
     let ctx = ctx!(); // create a context
     let mut sh = hashtable!()?; // create a hashtable
     hashtable_set_raw!(ctx, sh); // set raw mode

     // generate our own hash manually
     let mut hasher = DefaultHasher::new();
     (b"hi").hash(&mut hasher);
     let hash = hasher.finish();

     // insert
     sh.insert_raw(ctx, b"hi", usize!(hash), b"ok")?;

     // call get_raw to lookup our entry
     let slab = sh.get_raw(ctx, b"hi", usize!(hash))?.unwrap();

     // key = 104/105 (hi), value = 111/107 (ok)
     assert_eq!(
         slab.get()[0..36],
         [
             255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255, 255,
             0, 0, 0, 0, 0, 0, 0, 2, 104, 105, 0, 0, 0, 0, 0, 0, 0, 2, 111, 107
         ]
     );

     Ok(())
 }

fn remove_raw(
&mut self,
ctx: &mut Context,
key: &[u8],
hash: usize
) -> Result<bool, Error>

Remove raw data from the crate::StaticHashtable. On success returns Ok(bool), on failure returns bmw_err::Error. If the element is found, true is returned, otherwise, false.

Examples

 use bmw_util::{hashtable, ctx, hashtable_set_raw};
 use bmw_err::Error;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use bmw_log::usize;

 fn raw_ex() -> Result<(), Error> {
     let ctx = ctx!(); // create a context
     let mut sh = hashtable!()?; // create a hashtable
     hashtable_set_raw!(ctx, sh); // set raw mode

     // generate our own hash manually
     let mut hasher = DefaultHasher::new();
     (b"hi").hash(&mut hasher);
     let hash = hasher.finish();

     // insert
     sh.insert_raw(ctx, b"hi", usize!(hash), b"ok")?;

     sh.remove_raw(ctx, b"hi", usize!(hash))?;
     // call get_raw to lookup our entry which is no longer in the table
     let slab = sh.get_raw(ctx, b"hi", usize!(hash))?;
     assert!(slab.is_none());

     Ok(())
 }

fn iter_raw<'b>(&'b self, ctx: &mut Context) -> RawHashtableIterator<'b>

Returns an iterator for iterating through raw data in this crate::StaticHashtable. This is distinct from the iterator that is implemented as crate::StaticHashtable’s std::iter::IntoIterator trait which returns the serialized and not raw data. An understanding of the structure of the raw data in the crate::StaticHashtable’s slabs is necessary to use this function.

Examples

 use bmw_err::*;
 use bmw_util::{hashtable, hashtable_set_raw, ctx};
 use bmw_log::*;

 info!();

 fn test() -> Result<(), Error> {
     // create a context
     let ctx = ctx!();
     // create a hashtable
     let mut hashtable = hashtable!()?;
     // need to do a empty insert so that the types can be inferred. From here on we use raw
     // operations. The hashtable_set_raw macro handles this.
     hashtable_set_raw!(ctx, hashtable);
     let key = [0u8, 1u8, 123u8];
     let value = [10u8];
     let hash = 123usize;
     hashtable.insert_raw(ctx, &key, hash, &value)?;

     let key = [1u8, 1u8, 125u8];
     let value = [14u8];
     let hash = 125usize;
     hashtable.insert_raw(ctx, &key, hash, &value)?;

     let mut count = 0;
     for slab in hashtable.iter_raw(ctx) {
         info!("slab={:?}", slab.get())?;
         count += 1;
     }

     assert_eq!(count, 3);

     Ok(())
 }

fn size(&self, ctx: &mut Context) -> usize

Return the current size of this crate::StaticHashtable.

fn first_entry(&self, ctx: &mut Context) -> usize

Return the slab id of the first entry in this crate::StaticHashtable. note: this function is mostly used internally by iterators.

fn slab<'b>(&'b self, ctx: &mut Context, id: usize) -> Result<Slab<'b>, Error>

Return this crate::Slab associated with this id. note: this function is mostly used internally by iterators.

fn read_kv(&self, ctx: &mut Context, slab_id: usize) -> Result<(K, V), Error>

Return the key value pair associated with this slab id in serialized form.

fn get_array(&self, ctx: &mut Context) -> &Vec<usize>

Return an immutable reference to the underlying array backing this crate::StaticHashtable. note: this function is mostly used internally by iterators.

fn clear(&mut self, ctx: &mut Context) -> Result<(), Error>

Clear all elements in this crate::StaticHashtable and free any slabs that were associated with this table.