Trait bmw_util::ThreadPool

pub trait ThreadPool<T, OnPanic>
where
    OnPanic: FnMut(u128, Box<dyn Any + Send>) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,
{
    // Required methods
    fn execute<F>(&self, f: F, id: u128) -> Result<ThreadPoolHandle<T>, Error>
       where F: Future<Output = Result<T, Error>> + Send + 'static;
    fn start(&mut self) -> Result<(), Error>;
    fn stop(&mut self) -> Result<(), Error>;
    fn size(&self) -> Result<usize, Error>;
    fn stopper(&self) -> Result<ThreadPoolStopper, Error>;
    fn executor(&self) -> Result<ThreadPoolExecutor<T>, Error>
       where T: Send + Sync;
    fn set_on_panic(&mut self, on_panic: OnPanic) -> Result<(), Error>;
}

A thread pool. See crate::thread_pool for working examples.

Required Methods

fn execute<F>(&self, f: F, id: u128) -> Result<ThreadPoolHandle<T>, Error>
where F: Future<Output = Result<T, Error>> + Send + 'static,

Execute a task in the thread pool. This task will run to completion on the first available thread in the pool. The return value crate::ThreadPoolHandle which can be used to get the id of the task sent to the thread pool or to block on.

fn start(&mut self) -> Result<(), Error>

Start the pool. If macros are used, this call is unnecessary.

fn stop(&mut self) -> Result<(), Error>

Stop the thread pool. This function will ensure no new tasks are processed in the ThreadPool and that the threads will be stopped once they become idle again. It however, does not ensure that any tasks currently running in the thread pool are stopped immediately. That is the responsibility of the user.

fn size(&self) -> Result<usize, Error>

Returns the current size of the thread pool which will be between the configured maximum and minimum size.

fn stopper(&self) -> Result<ThreadPoolStopper, Error>

Get the crate::ThreadPoolStopper for this thread pool.

fn executor(&self) -> Result<ThreadPoolExecutor<T>, Error>
where T: Send + Sync,

Get the crate::ThreadPoolExecutor for this thread pool.

fn set_on_panic(&mut self, on_panic: OnPanic) -> Result<(), Error>

Set an on panic handler for this thread pool

Object Safety

This trait is not object safe.