Crate bmw_util

Macros

array

The crate::array! macro builds an crate::Array. The macro takes the following parameters:

array_list

The crate::array_list macro builds an crate::ArrayList in the form of a impl SortableList. The macro takes the following parameters:

array_list_box

This macro is identical to crate::array_list except that the value is returned in a box. To be exact, the return value is Box<dyn SortableList>. The boxed version can then be used to store in structs and enums. See crate::array_list for more details and an example.

array_list_sync

sync version of crate::array_list.

array_list_sync_box

sync box version of crate::array_list.

block_on

Macro used to block until a thread pool has completed the task. See crate::ThreadPool for working examples.

execute

Macro used to execute tasks in a thread pool. See crate::ThreadPool for working examples.

global_slab_allocator

The global_slab_allocator macro initializes the global thread local slab allocator for the thread that it is executed in. It takes the following parameters:

hashset

The crate::hashset macro builds a crate::Hashset with the specified configuration and optionally the specified crate::SlabAllocator. The macro accepts the following parameters:

hashset_box

The crate::hashset_box macro is the same as the crate::hashset macro except that the hashset is returned in a box. See crate::hashset.

hashset_sync

The hashset_sync macro is the same as crate::hashset except that the returned Hashset implements Send and Sync and can be safely passed through threads. See crate::hashtable_sync for further details.

hashset_sync_box

The hashset_sync_box macro is the boxed version of the crate::hashset_sync macro. It is the same except that the returned crate::Hashset is in a Box so it can be added to structs and enums.

hashtable

The crate::hashtable macro builds a crate::Hashtable with the specified configuration and optionally the specified crate::SlabAllocator. The macro accepts the following parameters:

hashtable_box

The crate::hashtable_box macro builds a crate::Hashtable with the specified configuration and optionally the specified crate::SlabAllocator. The only difference between this macro and the crate::hashtable macro is that the returned hashtable is inserted into a Box. Specifically, the return type is a Box<dyn Hashtable>. See crate::hashtable for further details.

hashtable_sync

The difference between this macro and the crate::hashtable macro is that the returned crate::Hashtable implements the Send and Sync traits and is thread safe. With this hashtable you cannot specify a crate::SlabAllocator because they use std::cell::RefCell which is not thread safe. That is also why this macro returns an error if ConfigOption::Slabs is specified. The parameters for this macro are:

hashtable_sync_box

This macro is the same as hashtable_sync except that the returned hashtable is in a Box. This macro can be used if the sync hashtable needs to be placed in a struct or an enum. See crate::hashtable and crate::hashtable_sync for further details.

list

The list macro is used to create lists. This macro uses the global slab allocator. To use a specified slab allocator, see crate::UtilBuilder::build_list. It has the same syntax as the std::vec! macro. Note that this macro and the builder function both return an implementation of the crate::SortableList trait.

list_append

Append list2 to list1.

list_box

This is the boxed version of list. The returned value is Box<dyn SortableList>. Otherwise, this macro is identical to crate::list.

list_eq

Compares equality of list1 and list2.

list_sync

Like crate::hashtable_sync and crate::hashset_sync list has a ‘sync’ version. See those macros for more details and see the crate for an example of the sync version of a hashtable. Just as in that example the list can be put into a crate::lock! or crate::lock_box and passed between threads.

list_sync_box

Box version of the crate::list_sync macro.

lock

Macro to get a crate::Lock. Internally, the parameter passed in is wrapped in an Arc<Rwlock<T>> wrapper that can be used to obtain read/write locks around any data structure.

lock_box

The same as lock except that the value returned is in a Box<dyn LockBox<T>> structure. See crate::LockBox for a working example.

pattern

The pattern macro builds a crate::Pattern which is used by the crate::SearchTrie. The pattern macro takes the following parameters:

queue

This macro creates a crate::Queue. The parameters are

queue_box

This is the box version of crate::queue. It is identical other than the returned value is in a box (Box<dyn Queue>).

queue_sync

This is the sync version of crate::queue. It is identical other than the returned value is with Sync/Send traits implemented.

queue_sync_box

This is the box version of crate::queue. It is identical other than the returned value is in a box (Box<dyn Queue>) and Send/Sync traits implemented.

rlock

macro to call rlock and guard function on a crate::LockBox at the same time. Note that this only allows a single access to the variable. If more than one operation needs to be done, this macro should not be used.

search_trie

The search_trie macro builds a crate::SearchTrie which can be used to match multiple patterns for a given text in a performant way. The search_trie macro takes the following parameters:

slab_allocator

The slab_allocator macro initializes a slab allocator with the specified parameters. It takes the following parameters:

stack

This macro creates a crate::Stack. The parameters are

stack_box

This is the box version of crate::stack. It is identical other than the returned value is in a box (Box<dyn Stack>).

stack_sync

sync version of crate::stack.

stack_sync_box

box version of crate::stack.

thread_pool

Macro used to configure/build a thread pool. See crate::ThreadPool for working examples.

tmatch

wlock

macro to call wlock and guard function on a crate::LockBox at the same time. Note that this only allows a single access to the variable. If more than one operation needs to be done, this macro should not be used.

Structs

Array

ArrayList

HashsetConfig

The configuration struct for a Hashset. This struct is passed into the crate::UtilBuilder::build_hashset function. The std::default::Default trait is implemented for this trait.

HashsetIterator

An iterator for the crate::Hashset.

HashtableConfig

The configuration struct for a Hashtable. This struct is passed into the crate::UtilBuilder::build_hashtable function. The std::default::Default trait is implemented for this trait.

HashtableIterator

An iterator for the crate::Hashtable.

ListConfig

Configuration for Lists currently there are no parameters, but it is still used to stay consistent with crate::Hashtable and crate::Hashset.

ListIterator

An iterator for the crate::List.

Match

Pattern

RwLockReadGuardWrapper

Wrapper around the std::sync::RwLockReadGuard.

RwLockWriteGuardWrapper

Wrapper around the std::sync::RwLockWriteGuard.

Slab

Struct that is used as a immutable reference to data in a slab. See crate::SlabAllocator for further details.

SlabAllocatorConfig

SlabMut

Struct that is used as a mutable reference to data in a slab. See crate::SlabAllocator for further details.

SlabReader

Utility to read from slabs using the [bmw_ser::Reader] trait.

SlabWriter

Utility to write to slabs using the [bmw_ser::Writer] trait.

ThreadPoolConfig

The configuration struct for a crate::ThreadPool. This struct is passed into the crate::UtilBuilder::build_thread_pool function or the crate::thread_pool macro. The std::default::Default trait is implemented for this trait. Also see ConfigOption for details on configuring via macro.

ThreadPoolExecutor

Struct that can be used to execute tasks in the thread pool. Mainly needed for passing the execution functionality to structs/threads.

ThreadPoolStopper

Struct that can be used to stop the thread pool. Note the limitations in crate::ThreadPoolStopper::stop.

UtilBuilder

Enums

PatternParam

PoolResult

SearchParam

Constants

RAND_CONTEXT

Traits

Hashset

Hashtable

List

Lock

LockBox

Queue

SearchTrie

SlabAllocator

SortableList

Stack

ThreadPool

Functions

canonicalize_base_path

Utility to canonicalize a path relative to a base directory (used in http and rustlet)

lock_box_from_usize

Rebuild a crate::LockBox from te usize which is returned from the crate::LockBox::danger_to_usize function.

random_bytes

random_u32

random_u64

random_u128

slice_to_u32

Utility to convert an arbitrary length slice (up to 8 bytes) to a u32.

slice_to_u64

Utility to convert an arbitrary length slice (up to 8 bytes) to a u64.

slice_to_u128

Utility to convert an arbitrary length slice (up to 16 bytes) to a u128.

slice_to_usize

Utility to convert an arbitrary length slice (up to 8 bytes) to a usize.

u32_to_slice

Utility to convert a u32 to an arbitrary length slice (up to 4 bytes).

u64_to_slice

Utility to convert a u64 to an arbitrary length slice (up to 8 bytes).

u128_to_slice

Utility to convert a u128 to an arbitrary length slice (up to 16 bytes).

usize_to_slice

Utility to convert a usize to an arbitrary length slice (up to 8 bytes).