Crate bmw_util

The util crate provides the data structures that are used in bmw. The data structures implemented include crate::Hashtable, crate::Hashset, crate::List, crate::Array, crate::ArrayList, crate::Stack, crate::Queue, and crate::SuffixTree. Additionally a crate::ThreadPool implementation is provided. Each data structure can be instantiated in a impl and Box form using the crate::Builder or through macros. The impls completely stack based and the box forms are Box<dyn ..>’s that can be stored in other structs and Enums.

Motivation

The advantage of these implementations is that they do not allocate memory on the heap after initialization of the data structure. So, we can create a crate::hashtable, crate::List or a crate::hashset and do millions of operations and no heap memory will be allocated or deallocated. Dynamic heap allocations that are long-lived can cause substantial problems like slowness and memory fragmentation and even system crashes and these data structures are intended to alleviate those problems. The core::ops::Drop trait is also implemented so all slabs used by the data structure are freed when the data structure goes out of scope.

Performance

The hashtable/set are not as fast as the native Rust data structures because they require serialization and deserialization of the entries on each operation. However, the peformance is at least in the ballpark of the standard data structures. The array, arraylist, queue, and stack are faster for insert, slower for initialization and about the same for iteration and drop. A performance tool is included in the project in the etc directory ds_perf.

Below is the output of some runs on a linux machine.

 $ ./target/release/perf  --help
 ds_perf 1.0
 37 Miners, LLC

 USAGE:
    perf [FLAGS]

 FLAGS:
     --array           run tests for array
     --array_string    run tests for array with strings
     --arraylist       run tests for array list
     --hashmap         run tests for standard rust library hashmap
     --hashtable       run tests for hashtable
 -h, --help            Prints help information
 -V, --version         Prints version information
     --vec             run tests for standard rust library vec
     --vec_string      run tests for vec with strings

 $ ./target/release/perf  --array_string
 [2022-09-01 13:02:41.974]: Starting ds_perf
 [2022-09-01 13:02:41.974]: Testing array string
 [2022-09-01 13:02:41.974]: array init: alloc: 240,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 240,000, elapsed: 180.336µs
 [2022-09-01 13:02:41.975]: array insert: alloc: 110,000, dealloc: 0, alloc_qty: 10,000, dealloc_qty: 0, delta: 110,000, elapsed: 332.891µs
 [2022-09-01 13:02:41.975]: array iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 27ns
 [2022-09-01 13:02:41.975]: array drop: alloc: 0, dealloc: 350,000, alloc_qty: 0, dealloc_qty: 10,001, delta: -350,000, elapsed: 121.148µs

 $ ./target/release/perf  --vec_string
 [2022-09-01 13:02:38.151]: Starting ds_perf
 [2022-09-01 13:02:38.151]: Testing vec string
 [2022-09-01 13:02:38.151]: vec init: alloc: 240,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 240,000, elapsed: 177.351µs
 [2022-09-01 13:02:38.151]: vec insert: alloc: 110,000, dealloc: 0, alloc_qty: 10,000, dealloc_qty: 0, delta: 110,000, elapsed: 324.671µs
 [2022-09-01 13:02:38.151]: vec iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 40ns
 [2022-09-01 13:02:38.152]: vec drop: alloc: 0, dealloc: 350,000, alloc_qty: 0, dealloc_qty: 10,001, delta: -350,000, elapsed: 121.868µs

 $ ./target/release/perf  --array
 [2022-09-01 13:00:22.190]: Starting ds_perf
 [2022-09-01 13:00:22.190]: Testing array
 [2022-09-01 13:00:22.190]: array init: alloc: 40,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 40,000, elapsed: 27.454µs
 [2022-09-01 13:00:22.190]: array insert: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 999ns
 [2022-09-01 13:00:22.190]: array iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 29ns
 [2022-09-01 13:00:22.190]: array drop: alloc: 0, dealloc: 40,000, alloc_qty: 0, dealloc_qty: 1, delta: -40,000, elapsed: 121ns

 $ ./target/release/perf  --vec
 [2022-09-01 13:00:27.820]: Starting ds_perf
 [2022-09-01 13:00:27.820]: testing vec
 [2022-09-01 13:00:27.820]: vec init: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 120ns
 [2022-09-01 13:00:27.820]: vec insert: alloc: 131,056, dealloc: 65,520, alloc_qty: 13, dealloc_qty: 12, delta: 65,536, elapsed: 48.554µs
 [2022-09-01 13:00:27.820]: vec iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 29ns
 [2022-09-01 13:00:27.820]: vec drop: alloc: 0, dealloc: 65,536, alloc_qty: 0, dealloc_qty: 1, delta: -65,536, elapsed: 5.096µs

 $ ./target/release/perf  --hashtable
 [2022-09-01 13:07:30.861]: Starting ds_perf
 [2022-09-01 13:07:30.861]: Testing hashtable
 [2022-09-01 13:07:30.861]: hashtable init: alloc: 100,096, dealloc: 96, alloc_qty: 2, dealloc_qty: 1, delta: 100,000, elapsed: 55.208µs
 [2022-09-01 13:07:30.863]: hashtable insert: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 2.296961ms
 [2022-09-01 13:07:30.865]: hashtable get: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 1.462402ms
 [2022-09-01 13:07:30.865]: hashtable drop: alloc: 100,000, dealloc: 200,000, alloc_qty: 1, dealloc_qty: 2, delta: -100,000, elapsed: 512.423µs

 $ ./target/release/perf  --hashmap
 [2022-09-01 13:07:39.269]: Starting ds_perf
 [2022-09-01 13:07:39.269]: Testing hashmap
 [2022-09-01 13:07:39.269]: hashmap init: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 69ns
 [2022-09-01 13:07:39.270]: hashmap insert: alloc: 688,252, dealloc: 344,172, alloc_qty: 26, dealloc_qty: 24, delta: 344,080, elapsed: 611.785µs
 [2022-09-01 13:07:39.270]: hashmap get: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 199.428µs
 [2022-09-01 13:07:39.270]: hashmap drop: alloc: 0, dealloc: 344,080, alloc_qty: 0, dealloc_qty: 2, delta: -344,080, elapsed: 25.252µs

 $ ./target/release/perf  --arraylist
 [2022-09-01 13:08:08.434]: Starting ds_perf
 [2022-09-01 13:08:08.435]: testing arraylist
 [2022-09-01 13:08:08.435]: arraylist init: alloc: 40,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 40,000, elapsed: 27.827µs
 [2022-09-01 13:08:08.435]: arraylist insert: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 18.963µs
 [2022-09-01 13:08:08.435]: arraylist iter: alloc: 24, dealloc: 24, alloc_qty: 1, dealloc_qty: 1, delta: 0, elapsed: 20.341µs
 [2022-09-01 13:08:08.435]: arraylist drop: alloc: 1,856, dealloc: 41,856, alloc_qty: 27, dealloc_qty: 28, delta: -40,000, elapsed: 110ns

So, as seen, the performance is somewhat comparable and keep in mind, that these operations are being done in the sub millisecond time frame, so it’s very unlikely that this will ultimately impact the performance of the application unless it’s an already very optimized application that needs to be very scalable.

Using bmw in your project

To use the crates in bmw in your project, add the following to your Cargo.toml

Use cases

The main use case for these data structures is in server applications to avoid making dynamic heap allocations as runtime, but they also offer some other interesting properties. For instance, with the standard rust collections, the entries in the hashmap are just references so they must stay in scope while they are in the hashmap. With this implementation, that is not required. The inserted items can be dropped and they will remain in the hashtable/hashset. Also, crate::Hashtable and crate::Hashset both implement the bmw_ser::Serializable trait so they can be sent from one part of an app to another or even sent over the network.

Examples

Re-exports

pub use crate::types::ConfigOption::*;

pub use crate::types::PatternParam::*;

pub use crate::types::SuffixParam::*;

Macros

array

array_list

array_list_box

block_on

Macro used to block until a thread pool has completed the task. See crate::ThreadPool for working examples.

execute

Macro used to execute tasks in a thread pool. See crate::ThreadPool for working examples.

hashset

hashset_box

hashset_config

hashset_sync

hashset_sync_box

hashset_sync_config

hashtable

hashtable_box

hashtable_config

hashtable_sync

hashtable_sync_box

hashtable_sync_config

init_slab_allocator

list

list_append

list_box

list_eq

list_sync

list_sync_box

pattern

queue

queue_box

slab_allocator

stack

stack_box

suffix_tree

thread_pool

Macro used to configure/build a thread pool. See crate::ThreadPool for working examples.

Structs

Array

ArrayIterator

ArrayList

BinReader

Utility wrapper for an underlying byte Reader. Defines higher level methods to write numbers, byte vectors, hashes, etc.

BinWriter

Utility wrapper for an underlying byte Writer. Defines higher level methods to write numbers, byte vectors, hashes, etc.

Builder

HashsetConfig

The configuration struct for a Hashset. This struct is passed into the crate::Builder::build_hashset function. The std::default::Default trait is implemented for this trait.

HashsetIterator

HashtableConfig

The configuration struct for a Hashtable. This struct is passed into the crate::Builder::build_hashtable function. The std::default::Default trait is implemented for this trait.

HashtableIterator

ListConfig

ListIterator

Match

Pattern

Slab

Struct that is used as a immutable refernce to data in a slab. See crate::SlabAllocator for further details.

SlabAllocatorConfig

Slab Allocator configuration struct. This struct is the input to the crate::SlabAllocator::init function. The two parameters are slab_size which is the size of the slabs in bytes allocated by this crate::SlabAllocator and slab_count which is the number of slabs that can be allocated by this crate::SlabAllocator.

SlabMut

Struct that is used as a mutable refernce to data in a slab. See crate::SlabAllocator for further details.

SlabReader

SlabWriter

ThreadPoolConfig

The configuration struct for a crate::ThreadPool. This struct is passed into the crate::Builder::build_thread_pool function or the crate::thread_pool macro. The std::default::Default trait is implemented for this trait. Also see crate::ConfigOption for details on configuring via macro.

Enums

ConfigOption

Configuration options used throughout this crate via macro.

PatternParam

PoolResult

The result returned from a call to crate::ThreadPool::execute. This is similar to std::result::Result except that it implements std::marker::Send and std::marker::Sync so that it can be passed through threads. Also a type crate::PoolResult::Panic is returned if a thread panic occurs in the thread pool.

SuffixParam

Traits

Hashset

Hashtable

List

Queue

Reader

Reader trait used for deserializing data.

Serializable

This is the trait used by all data structures to serialize and deserialize data. Anthing stored in them must implement this trait. Commonly needed implementations are built in the ser module in this crate. These include Vec, String, integer types among other things.

SlabAllocator

This trait defines the public interface to the crate::SlabAllocator. The slab allocator is used by the other data structures in this crate to avoid dynamic heap allocations. By itself, the slab allocator is fairly simple. It only allocates and frees slabs. crate::SlabAllocator::get and crate::SlabAllocator::get_mut are also provided to obtain immutable and mutable references to a slab respectively. They only contain references to the data and not copies.

SortableList

Stack

SuffixTree

ThreadPool

This trait defines the public interface to the ThreadPool. A pool can be configured via the crate::ThreadPoolConfig struct. The thread pool should be accessed through the macros under normal circumstances. See crate::thread_pool, crate::execute and crate::block_on for additional details. The thread pool can be passed through threads via a bmw_log::Lock or bmw_log::LockBox so a single thread pool can service multiple worker threads. See examples below.

Writer

Writer trait used to serializing data.

Functions

deserialize

Deserializes a Serializable from any std::io::Read implementation.

serialize

Serializes a Serializable into any std::io::Write implementation.