Crate bmw_util

The util crate provides the data structures that are used in bmw. The data structures implemented include crate::Hashtable, crate::Hashset, crate::List, crate::Array, crate::ArrayList, crate::Stack, crate::Queue, and crate::SuffixTree. Additionally a crate::ThreadPool implementation is provided. Each data structure can be instantiated in a impl and Box form using the crate::Builder or through macros. The impls completely stack based and the box forms are Box<dyn ..>’s that can be stored in other structs and Enums. While the boxed versions do store data on the heap, it is only pointers and the vast majority of the data, in either form, is stored in pre-allocated slabs. The array based structures do use the heap, they use std::vec::Vec as the underlying storage mechanism, but they only allocate heap memory when they are created and none afterwords.

Motivation

The advantage of these implementations is that they do not allocate memory on the heap after initialization of the data structure. So, we can create a crate::hashtable, crate::List or a crate::hashset and once created, do millions of operations and no heap memory will be allocated or deallocated. Dynamic heap allocations that are long-lived can cause substantial problems like slowness and memory fragmentation and even system crashes and these data structures are intended to alleviate those problems. The core::ops::Drop trait is also implemented so all slabs used by the data structure are freed when the data structure goes out of scope.

Performance

The hashtable/set are not as fast as the native Rust data structures because they require serialization and deserialization of the entries on each operation. However, the performance is at least in the ballpark of the standard data structures. The array, arraylist, queue, and stack are faster for insert, slower for initialization and about the same for iteration and drop. A performance tool is included in the project in the etc directory ds_perf.

Below is the output of some runs on a linux machine.

 $ ./target/release/perf  --help
 ds_perf 1.0
 37 Miners, LLC

 USAGE:
    perf [FLAGS]

 FLAGS:
     --array           run tests for array
     --array_string    run tests for array with strings
     --arraylist       run tests for array list
     --hashmap         run tests for standard rust library hashmap
     --hashtable       run tests for hashtable
 -h, --help            Prints help information
 -V, --version         Prints version information
     --vec             run tests for standard rust library vec
     --vec_string      run tests for vec with strings

 $ ./target/release/perf  --array_string
 [2022-09-01 13:02:41.974]: Starting ds_perf
 [2022-09-01 13:02:41.974]: Testing array string
 [2022-09-01 13:02:41.974]: array init: alloc: 240,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 240,000, elapsed: 180.336µs
 [2022-09-01 13:02:41.975]: array insert: alloc: 110,000, dealloc: 0, alloc_qty: 10,000, dealloc_qty: 0, delta: 110,000, elapsed: 332.891µs
 [2022-09-01 13:02:41.975]: array iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 27ns
 [2022-09-01 13:02:41.975]: array drop: alloc: 0, dealloc: 350,000, alloc_qty: 0, dealloc_qty: 10,001, delta: -350,000, elapsed: 121.148µs

 $ ./target/release/perf  --vec_string
 [2022-09-01 13:02:38.151]: Starting ds_perf
 [2022-09-01 13:02:38.151]: Testing vec string
 [2022-09-01 13:02:38.151]: vec init: alloc: 240,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 240,000, elapsed: 177.351µs
 [2022-09-01 13:02:38.151]: vec insert: alloc: 110,000, dealloc: 0, alloc_qty: 10,000, dealloc_qty: 0, delta: 110,000, elapsed: 324.671µs
 [2022-09-01 13:02:38.151]: vec iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 40ns
 [2022-09-01 13:02:38.152]: vec drop: alloc: 0, dealloc: 350,000, alloc_qty: 0, dealloc_qty: 10,001, delta: -350,000, elapsed: 121.868µs

 $ ./target/release/perf  --array
 [2022-09-01 13:00:22.190]: Starting ds_perf
 [2022-09-01 13:00:22.190]: Testing array
 [2022-09-01 13:00:22.190]: array init: alloc: 40,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 40,000, elapsed: 27.454µs
 [2022-09-01 13:00:22.190]: array insert: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 999ns
 [2022-09-01 13:00:22.190]: array iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 29ns
 [2022-09-01 13:00:22.190]: array drop: alloc: 0, dealloc: 40,000, alloc_qty: 0, dealloc_qty: 1, delta: -40,000, elapsed: 121ns

 $ ./target/release/perf  --vec
 [2022-09-01 13:00:27.820]: Starting ds_perf
 [2022-09-01 13:00:27.820]: testing vec
 [2022-09-01 13:00:27.820]: vec init: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 120ns
 [2022-09-01 13:00:27.820]: vec insert: alloc: 131,056, dealloc: 65,520, alloc_qty: 13, dealloc_qty: 12, delta: 65,536, elapsed: 48.554µs
 [2022-09-01 13:00:27.820]: vec iter: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 29ns
 [2022-09-01 13:00:27.820]: vec drop: alloc: 0, dealloc: 65,536, alloc_qty: 0, dealloc_qty: 1, delta: -65,536, elapsed: 5.096µs

 $ ./target/release/perf  --hashtable
 [2022-09-01 13:07:30.861]: Starting ds_perf
 [2022-09-01 13:07:30.861]: Testing hashtable
 [2022-09-01 13:07:30.861]: hashtable init: alloc: 100,096, dealloc: 96, alloc_qty: 2, dealloc_qty: 1, delta: 100,000, elapsed: 55.208µs
 [2022-09-01 13:07:30.863]: hashtable insert: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 2.296961ms
 [2022-09-01 13:07:30.865]: hashtable get: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 1.462402ms
 [2022-09-01 13:07:30.865]: hashtable drop: alloc: 100,000, dealloc: 200,000, alloc_qty: 1, dealloc_qty: 2, delta: -100,000, elapsed: 512.423µs

 $ ./target/release/perf  --hashmap
 [2022-09-01 13:07:39.269]: Starting ds_perf
 [2022-09-01 13:07:39.269]: Testing hashmap
 [2022-09-01 13:07:39.269]: hashmap init: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 69ns
 [2022-09-01 13:07:39.270]: hashmap insert: alloc: 688,252, dealloc: 344,172, alloc_qty: 26, dealloc_qty: 24, delta: 344,080, elapsed: 611.785µs
 [2022-09-01 13:07:39.270]: hashmap get: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 199.428µs
 [2022-09-01 13:07:39.270]: hashmap drop: alloc: 0, dealloc: 344,080, alloc_qty: 0, dealloc_qty: 2, delta: -344,080, elapsed: 25.252µs

 $ ./target/release/perf  --arraylist
 [2022-09-01 13:08:08.434]: Starting ds_perf
 [2022-09-01 13:08:08.435]: testing arraylist
 [2022-09-01 13:08:08.435]: arraylist init: alloc: 40,000, dealloc: 0, alloc_qty: 1, dealloc_qty: 0, delta: 40,000, elapsed: 27.827µs
 [2022-09-01 13:08:08.435]: arraylist insert: alloc: 0, dealloc: 0, alloc_qty: 0, dealloc_qty: 0, delta: 0, elapsed: 18.963µs
 [2022-09-01 13:08:08.435]: arraylist iter: alloc: 24, dealloc: 24, alloc_qty: 1, dealloc_qty: 1, delta: 0, elapsed: 20.341µs
 [2022-09-01 13:08:08.435]: arraylist drop: alloc: 1,856, dealloc: 41,856, alloc_qty: 27, dealloc_qty: 28, delta: -40,000, elapsed: 110ns

So, as seen, the performance is somewhat comparable and keep in mind, that these operations are being done in the sub millisecond time frame, so it’s very unlikely that this will ultimately impact the performance of the application unless it’s an already very optimized application that needs to be very scalable.

Using bmw_util (and other crates) in your project

To use the crates in bmw in your project, add the following to your Cargo.toml:

 bmw_util   = { git = "https://github.com/37miners/bmw"  }

Optionally, you may wish to use the other associated crates:

 bmw_err    = { git = "https://github.com/37miners/bmw"  }
 bmw_log    = { git = "https://github.com/37miners/bmw"  }
 bmw_derive = { git = "https://github.com/37miners/bmw"  }

The linux dependencies can be installed with the following commands on ubuntu:

 $ sudo apt-get update -yqq
 $ sudo apt-get install -yqq --no-install-recommends libncursesw5-dev libssl-dev

The macos dependencies can be installed with the following commands

$ brew install llvm

The windows dependencies can be installed with the following commands

$ choco install -y llvm

BMW is tested with the latest version of rust. Please ensure to update it.

Use cases

The main use case for these data structures is in server applications to avoid making dynamic heap allocations as runtime, but they also offer some other interesting properties. For instance, with the standard rust collections, the entries in the hashmap are just references so they must stay in scope while they are in the hashmap. With this implementation, that is not required. The inserted items can be dropped and they will remain in the hashtable/hashset. Also, crate::Hashtable and crate::Hashset both implement the bmw_ser::Serializable trait so they can be sent from one part of an app to another or even sent over the network.

Examples

Use of the slab allocator, hashtable, and hashset via macros.

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 info!();

 fn main() -> Result<(), Error> {
     // create a slab allocator to be used by our data structures
     let slabs = slab_allocator!(SlabSize(128), SlabCount(15_000))?;

     // instantiate a hashtable with specified MaxEntries and MaxLoadFactor
     let mut hashtable = hashtable!(MaxEntries(10_000), MaxLoadFactor(0.85), Slabs(&slabs))?;

     // insert a key/value pair (rust figures out the data types)
     hashtable.insert(&1, &2)?;

     // get the value for our key and assert it is correct
     let v = hashtable.get(&1)?;
     assert_eq!(v, Some(2));
     info!("value = {:?}", v)?;

     // instantiate a hashset that uses the same slab allocator. Since we don't specify
     // the MaxLoadFactor, the default value of 0.8 will be used for this hashset.
     let mut hashset = hashset!(MaxEntries(5_000), Slabs(&slabs))?;

     // insert a key
     hashset.insert(&"test".to_string())?;

     // assert that our key is found in the hashset
     assert!(hashset.contains(&"test".to_string())?);

     // do a negative assertion
     assert!(!hashset.contains(&"test2".to_string())?);

     info!("complete!")?;

     Ok(())
 }

Use of Sortable lists

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
     // for this example we will use the global slab allocator which is a
     // thread local slab allocator which we can configure via macro
     global_slab_allocator!(SlabSize(64), SlabCount(100_000))?;

     // create two lists (one linked and one array list).
     // Note that all lists created via macro are interoperable.
     let mut list1 = list![1u32, 2u32, 4u32, 5u32];
     let mut list2 = array_list!(10, &0u32)?;
     list2.push(5)?;
     list2.push(2)?;
     list2.push(4)?;
     list2.push(1)?;

     // sort the array list and assert it's equal to the other list
     list2.sort()?;
     assert!(list_eq!(list1, list2));

     // append list2 to list1 and assert the value
     list_append!(list1, list2);
     assert!(list_eq!(list1, list![1, 2, 4, 5, 1, 2, 4, 5]));

     Ok(())
 }

Arrays

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
     // instantiate an array of size 100 with the default values of 0usize
     let mut arr = array!(100, &0usize)?;

     // set the 10th element to 1
     arr[10] = 1;

     // set each item to the value of the iterator
     for i in 0..100 {
         arr[i] = i;
     }

     // assert the values
     for i in 0..100 {
         assert_eq!(arr[i], i);
     }

     Ok(())
 }

Thread pool and Locks

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
     // create a thread pool with the default settings
     let mut tp = thread_pool!()?;
     tp.set_on_panic(move |_id,_e| -> Result<(), Error> { Ok(()) })?;

     // create a lock initializing it's value to 0
     let x = lock!(0)?;
     // clone the lock (one for the thread pool, one for the local thread)
     let mut x_clone = x.clone();

     // execute in the thread pool
     let handle = execute!(tp, {
         // obtain the write lock for x
         let mut x = x_clone.wlock()?;
         // set the value to 1
         (**x.guard()) = 1;

         // return the value of 100
         Ok(100)
     })?;

     // block on the thread and assert the value returned is 100
     assert_eq!(block_on!(handle), PoolResult::Ok(100));

     // obtain the read lock for x
     let x = x.rlock()?;
     // assert the value of x is now 1
     assert_eq!(**x.guard(), 1);

     Ok(())
 }

Boxed and Sync versions

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 struct MyStruct {
     hashtable: Box<dyn Hashtable<u32, String>>,
 }

 fn main() -> Result<(), Error> {
     // create a slab allocator to be used by our data structures
     let slabs = slab_allocator!(SlabSize(128), SlabCount(15_000))?;

     {
     
         // instantiate a hashtable with specified MaxEntries and MaxLoadFactor
         let hashtable = hashtable_box!(
             MaxEntries(10_000),
             MaxLoadFactor(0.85),
             Slabs(&slabs)
         )?;

         let mut s = MyStruct {
             hashtable,
         };

         s.hashtable.insert(&1, &"something".to_string())?;
         s.hashtable.insert(&2, &"something else".to_string())?;

         assert_eq!(s.hashtable.get(&1)?, Some("something".to_string()));

     }

     // sync hashtable with default configuration
     let mut hashtable2 = hashtable_sync!()?;
     hashtable2.insert(&1, &10)?;
     hashtable2.insert(&2, &20)?;

     // create a thread pool with the default settings
     let mut tp = thread_pool!()?;
     tp.set_on_panic(move |_id,_e| -> Result<(), Error> { Ok(()) })?;

     // put the hashtable into a lock and clone it
     let mut lock = lock!(hashtable2)?;
     let lock_clone = lock.clone();

     // execute a task in the thread pool
     let handle = execute!(tp, {
         // obtain the write lock to the hashtable and insert an additional value
         let mut hashtable2 = lock.wlock()?;
         (**hashtable2.guard()).insert(&3, &30)?;
         Ok(())
     })?;

     // block on the task that is in the thread pool
     block_on!(handle);

     // obtain the read lock and assert all three values are in the hashtable
     let hashtable2 = lock_clone.rlock()?;
     assert_eq!((**hashtable2.guard()).get(&1)?, Some(10));
     assert_eq!((**hashtable2.guard()).get(&2)?, Some(20));
     assert_eq!((**hashtable2.guard()).get(&3)?, Some(30));

     Ok(())
 }

Suffix Tree

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
     // create an array of matches for the suffix tree to populate in tmatch
     let mut matches = [Builder::build_match_default(); 10];

     // create a suffix tree with two patterns with distinct Regexes and Ids
     // Setting a termination length (length to stop processing the text) of 100
     // and a max wild card length (the maximum length of any wildcards) of 50
     let mut suffix_tree = suffix_tree!(
         list![
             pattern!(Regex("abc"), Id(0))?,
             pattern!(Regex("def.*ok"), Id(1))?
         ],
         TerminationLength(100),
         MaxWildcardLength(50)
     )?;

     // run the matches and return the number of matches assert that it's
     // one for the abc match
     let match_count = suffix_tree.tmatch(b"abc", &mut matches)?;
     assert_eq!(match_count, 1);

     Ok(())
 }

Queues and Stacks

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
     // create a stack and a queue both with capacity of 1_000 items and &0
     // is the default value used to initialize the array
     let mut queue = queue!(1_000, &0)?;
     let mut stack = stack!(1_000, &0)?;

     // add three items to the queue
     queue.enqueue(1)?;
     queue.enqueue(2)?;
     queue.enqueue(3)?;

     // add the same three items to the stack
     stack.push(1)?;
     stack.push(2)?;
     stack.push(3)?;

     // dequeue/pop and assert that the queue/stack are appropriately returning the items
     assert_eq!(queue.dequeue(), Some(&1));
     assert_eq!(stack.pop(), Some(&3));

     // dequeue/pop and assert that the queue/stack are appropriately returning the items
     assert_eq!(queue.dequeue(), Some(&2));
     assert_eq!(stack.pop(), Some(&2));

     // peek at both the stack and queue (view value, but do not remove it from the queue)
     assert_eq!(queue.peek(), Some(&3));
     assert_eq!(stack.peek(), Some(&1));

     // dequeue/pop the last item
     assert_eq!(queue.dequeue(), Some(&3));
     assert_eq!(stack.pop(), Some(&1));

     // assert that the queue/stack are empty
     assert_eq!(queue.dequeue(), None);
     assert_eq!(stack.pop(), None);

     Ok(())
 }

Derive Serializable trait

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;
 use bmw_derive::*;

 info!();

 #[derive(Serializable, Clone, Debug, PartialEq)]
 struct MyStruct {
     id: u128,
     name: String,
     phone: Option<String>,
     age: u8,
 }

 fn main() -> Result<(), Error> {
     let s = MyStruct {
         id: 1234,
         name: "Hagrid".to_string(),
         phone: None,
         age: 54,
     };

     debug!("my struct = {:?}", s)?;

     let mut hashtable = hashtable!()?;

     hashtable.insert(&1, &s)?;

     let v = hashtable.get(&1)?;
     assert_eq!(v, Some(s));

     info!("value of record #1 is {:?}", v)?;
     
     Ok(())
 }

Re-exports

pub use crate::types::ConfigOption::*;

pub use crate::types::PatternParam::*;

pub use crate::types::SuffixParam::*;

Macros

array

The crate::array! macro builds an crate::Array. The macro takes the following parameters:

array_list

The crate::array_list macro builds an crate::ArrayList in the form of a impl SortableList. The macro takes the following parameters:

array_list_box

This macro is identical to crate::array_list except that the value is returned in a box. To be exact, the return value is Box<dyn SortableList>. The boxed version can then be used to store in structs and enums. See crate::array_list for more details and an example.

array_list_sync

sync version of crate::array_list.

array_list_sync_box

sync box version of crate::array_list.

block_on

Macro used to block until a thread pool has completed the task. See crate::ThreadPool for working examples.

execute

Macro used to execute tasks in a thread pool. See crate::ThreadPool for working examples.

global_slab_allocator

The global_slab_allocator macro initializes the global thread local slab allocator for the thread that it is executed in. It takes the following parameters:

hashset

The crate::hashset macro builds a crate::Hashset with the specified configuration and optionally the specified crate::SlabAllocator. The macro accepts the following parameters:

hashset_box

The crate::hashset_box macro is the same as the crate::hashset macro except that the hashset is returned in a box. See crate::hashset.

hashset_sync

The hashset_sync macro is the same as crate::hashset except that the returned Hashset implements Send and Sync and can be safely passed through threads. See crate::hashtable_sync for further details.

hashset_sync_box

The hashset_sync_box macro is the boxed version of the crate::hashset_sync macro. It is the same except that the returned crate::Hashset is in a Box so it can be added to structs and enums.

hashtable

The crate::hashtable macro builds a crate::Hashtable with the specified configuration and optionally the specified crate::SlabAllocator. The macro accepts the following parameters:

hashtable_box

The crate::hashtable_box macro builds a crate::Hashtable with the specified configuration and optionally the specified crate::SlabAllocator. The only difference between this macro and the crate::hashtable macro is that the returned hashtable is inserted into a Box. Specifically, the return type is a Box<dyn Hashtable>. See crate::hashtable for further details.

hashtable_sync

The difference between this macro and the crate::hashtable macro is that the returned crate::Hashtable implements the Send and Sync traits and is thread safe. With this hashtable you cannot specify a crate::SlabAllocator because they use std::cell::RefCell which is not thread safe. That is also why this macro returns an error if crate::ConfigOption::Slabs is specified. The parameters for this macro are:

hashtable_sync_box

This macro is the same as hashtable_sync except that the returned hashtable is in a Box. This macro can be used if the sync hashtable needs to be placed in a struct or an enum. See crate::hashtable and crate::hashtable_sync for further details.

list

The list macro is used to create lists. This macro uses the global slab allocator. To use a specified slab allocator, see crate::Builder::build_list. It has the same syntax as the std::vec! macro. Note that this macro and the builder function both return an implementation of the crate::SortableList trait.

list_append

Append list2 to list1.

list_box

This is the boxed version of list. The returned value is Box<dyn SortableList>. Otherwise, this macro is identical to crate::list.

list_eq

Compares equality of list1 and list2.

list_sync

Like crate::hashtable_sync and crate::hashset_sync list has a ‘sync’ version. See those macros for more details and see the crate for an example of the sync version of a hashtable. Just as in that example the list can be put into a crate::lock! or crate::lock_box and passed between threads.

list_sync_box

Box version of the crate::list_sync macro.

lock

Macro to get a crate::Lock. Internally, the parameter passed in is wrapped in an Arc<Rwlock> wrapper that can be used to obtain read/write locks around any data structure.

lock_box

The same as lock except that the value returned is in a Box<dyn LockBox> structure. See crate::LockBox for a working example.

pattern

The pattern macro builds a crate::Pattern which is used by the crate::SuffixTree. The pattern macro takes the following parameters:

queue

This macro creates a crate::Queue. The parameters are

queue_box

This is the box version of crate::queue. It is identical other than the returned value is in a box (Box<dyn Queue>).

queue_sync

This is the sync version of crate::queue. It is identical other than the returned value is with Sync/Send traits implemented.

queue_sync_box

This is the box version of crate::queue. It is identical other than the returned value is in a box (Box<dyn Queue>) and Send/Sync traits implemented.

slab_allocator

The slab_allocator macro initializes a slab allocator with the specified parameters. It takes the following parameters:

stack

This macro creates a crate::Stack. The parameters are

stack_box

This is the box version of crate::stack. It is identical other than the returned value is in a box (Box<dyn Stack>).

stack_sync

sync version of crate::stack.

stack_sync_box

box version of crate::stack.

suffix_tree

The suffix_tree macro builds a crate::SuffixTree which can be used to match multiple patterns for a given text in a performant way. The suffix_tree macro takes the following parameters:

thread_pool

Macro used to configure/build a thread pool. See crate::ThreadPool for working examples.

Structs

Array

The crate::Array is essentially a wrapper around std::vec::Vec. It is mainly used to prevent post initialization heap allocations. In many cases resizing/growing of the vec is not needed and adding this wrapper assures that they do not happen. There are use cases where growing is useful and in fact in the library we use Vec for the suffix tree, but in most cases where contiguous memory is used, the primary purpose is for indexing. That can be done with an array. So we add this wrapper.

ArrayIterator

An iterator for the crate::Array.

ArrayList

crate::ArrayList is an array based implementation of the crate::List and crate::SortableList trait. It uses the crate::Array as it’s underlying storage mechanism. In most cases it is more performant than the LinkedList implementation, but it does do a heap allocation when created. See the module level documentation for an example.

BinReader

Utility wrapper for an underlying byte Reader. Defines higher level methods to write numbers, byte vectors, hashes, etc.

BinWriter

Utility wrapper for an underlying byte Writer. Defines higher level methods to write numbers, byte vectors, hashes, etc.

Builder

The builder struct for the util library. All data structures are built through this interface. This is used by the macros as well.

HashsetConfig

The configuration struct for a Hashset. This struct is passed into the crate::Builder::build_hashset function. The std::default::Default trait is implemented for this trait.

HashsetIterator

An iterator for the crate::Hashset.

HashtableConfig

The configuration struct for a Hashtable. This struct is passed into the crate::Builder::build_hashtable function. The std::default::Default trait is implemented for this trait.

HashtableIterator

An iterator for the crate::Hashtable.

ListConfig

Configuration for Lists currently there are no parameters, but it is still used to stay consistent with crate::Hashtable and crate::Hashset.

ListIterator

An iterator for the crate::List.

Match

A match which is returned by the suffix tree. See crate::suffix_tree!.

Pattern

A pattern which is used with the suffix tree. See crate::suffix_tree!.

RwLockReadGuardWrapper

Wrapper around the std::sync::RwLockReadGuard.

RwLockWriteGuardWrapper

Wrapper around the std::sync::RwLockWriteGuard.

Slab

Struct that is used as a immutable reference to data in a slab. See crate::SlabAllocator for further details.

SlabAllocatorConfig

Slab Allocator configuration struct. This struct is the input to the crate::SlabAllocator::init function. The two parameters are slab_size which is the size of the slabs in bytes allocated by this crate::SlabAllocator and slab_count which is the number of slabs that can be allocated by this crate::SlabAllocator.

SlabMut

Struct that is used as a mutable reference to data in a slab. See crate::SlabAllocator for further details.

SlabReader

Utility to read from slabs using the bmw_ser::Reader trait.

SlabWriter

Utility to write to slabs using the bmw_ser::Writer trait.

ThreadPoolConfig

The configuration struct for a crate::ThreadPool. This struct is passed into the crate::Builder::build_thread_pool function or the crate::thread_pool macro. The std::default::Default trait is implemented for this trait. Also see crate::ConfigOption for details on configuring via macro.

ThreadPoolExecutor

Struct that can be used to execute tasks in the thread pool. Mainly needed for passing the execution functionality to structs/threads.

ThreadPoolStopper

Struct that can be used to stop the thread pool. Note the limitations in crate::ThreadPoolStopper::stop.

Enums

ConfigOption

Configuration options used throughout this crate via macro.

PatternParam

The enum used to define patterns. See crate::pattern for more info.

PoolResult

The result returned from a call to crate::ThreadPool::execute. This is similar to std::result::Result except that it implements std::marker::Send and std::marker::Sync so that it can be passed through threads. Also a type crate::PoolResult::Panic is returned if a thread panic occurs in the thread pool.

SuffixParam

The enum used to define a suffix_tree. See crate::suffix_tree! for more info.

Traits

Hashset

A slab allocated Hashset. Most of the implementation is shared with crate::Hashtable. See crate::Hashtable for a discussion of the slab layout. The difference is that as is the case with hashsets, there is no value.

Hashtable

A slab allocated hashtable. Data is stored in a crate::SlabAllocator defined by the user or using a global thread local slab allocator. All keys and values must implement the bmw_ser::Serializable trait which can be implemented with the bmw_derive::Serializable proc_macro.

List

A trait that defines a list. Both an array list and a linked list are implemented the default crate::list macro returns the linked list. The crate::array_list macro returns the array list. Both implement this trait.

Lock

Wrapper around the lock functionalities used by bmw in std::sync rust libraries. The main benefits are the simplified interface and the fact that if a thread attempts to obtain a lock twice, an error will be thrown instead of a thread panic. This is implemented through a thread local Hashset which keeps track of the guards used by the lock removes an entry for them when the guard is dropped.

LockBox

crate::LockBox is the same as crate::Lock except that it is possible to build The LockBox into a Box<dyn LockBox> structure so that it is object safe. It can then be cloned using DynClone.

Queue

This trait defines a queue. The implementation is a bounded queue. The queue uses an crate::Array as the underlying storage mechanism.

Reader

Reader trait used for deserializing data.

Serializable

This is the trait used by all data structures to serialize and deserialize data. Anything stored in them must implement this trait. Commonly needed implementations are built in the ser module in this crate. These include Vec, String, integer types among other things.

SlabAllocator

This trait defines the public interface to the crate::SlabAllocator. The slab allocator is used by the other data structures in this crate to avoid dynamic heap allocations. By itself, the slab allocator is fairly simple. It only allocates and frees slabs. crate::SlabAllocator::get and crate::SlabAllocator::get_mut are also provided to obtain immutable and mutable references to a slab respectively. They only contain references to the data and not copies.

SortableList

A trait that defines a sortable list. Both implementations implement this trait, although the linked list merely copies the data into an array list to sort. The array_list can natively sort with rust’s sort implementations using the functions associated with slice.

Stack

This trait defines a stack. The implementation is a bounded stack. The stack uses an crate::Array as the underlying storage mechanism.

SuffixTree

The suffix tree data structure. See crate::suffix_tree!.

ThreadPool

This trait defines the public interface to the ThreadPool. A pool can be configured via the crate::ThreadPoolConfig struct. The thread pool should be accessed through the macros under normal circumstances. See crate::thread_pool, crate::execute and crate::block_on for additional details. The thread pool can be passed through threads via a crate::Lock or crate::LockBox so a single thread pool can service multiple worker threads. See examples below.

Writer

Writer trait used to serializing data.

Functions

deserialize

Deserializes a Serializable from any std::io::Read implementation.

lock_box_from_usize

Rebuild a crate::LockBox from te usize which is returned from the crate::LockBox::danger_to_usize function.

serialize

Serializes a Serializable into any std::io::Write implementation.