Macro bmw_util::global_slab_allocator

macro_rules! global_slab_allocator {
    ( $( $config:expr ),* ) => { ... };
}

The global_slab_allocator macro initializes the global thread local slab allocator for the thread that it is executed in. It takes the following parameters:

SlabSize(usize) (optional) - the size in bytes of the slabs for this slab allocator. if not specified, the default value of 256 is used.

SlabCount(usize) (optional) - the number of slabs to allocate to the global slab allocator. If not specified, the default value of 40,960 is used.

Return

Return Ok(()) on success or [bmw_err::Error] on failure.

Errors

[bmw_err::ErrorKind::Configuration] - Is returned if a crate::ConfigOption other than crate::ConfigOption::SlabSize or crate::ConfigOption::SlabCount is specified.

[bmw_err::ErrorKind::IllegalState] - Is returned if the global thread local slab allocator has already been initialized for the thread that executes the macro. This can happen if the macro is called more than once or if a data structure that uses the global slab allocator is initialized and in turn initializes the global slab allocator with default values.

[bmw_err::ErrorKind::IllegalArgument] - Is returned if the SlabSize is 0 or the SlabCount is 0.

Examples

 use bmw_util::*;
 use bmw_err::Error;

 fn main() -> Result<(), Error> {
     global_slab_allocator!(SlabSize(128), SlabCount(1_000))?;

     // this will use the global slab allocator since we don't specify one
     let hashtable: Box<dyn Hashtable<u32, u32>> = hashtable_box!()?;

     // ...

     Ok(())
 }