Macro bmw_util::hashtable

macro_rules! hashtable {
    ( $( $config:expr ),* ) => { ... };
}

The crate::hashtable macro builds a crate::Hashtable with the specified configuration and optionally the specified crate::SlabAllocator. The macro accepts the following parameters:

MaxEntries(usize) (optional) - The maximum number of entries that can be in this hashtable at any given time. If not specified, the default value of 100_000 will be used.

MaxLoadFactor(usize) (optional) - The maximum load factor of the hashtable. The hashtable is array based hashtable and it has a fixed size. Once the load factor is reach, insertions will return an error. The hashtable uses linear probing to handle collisions. The max_load_factor makes sure no additional insertions occur at a given ratio of entries to capacity in the array. Note that MaxEntries can always be inserted, it’s the capacity of the array that becomes larger as this ratio goes down. If not specified, the default value is 0.8.

Slabs(Option<&Rc<RefCell>>) (optional) - An optional reference to a slab allocator to use with this crate::Hashtable. If not specified, the global slab allocator is used.

Returns

A Ok(impl Hashtable<K, V>) on success or a [bmw_err::Error] on failure.

Errors

[bmw_err::ErrorKind::Configuration] if anything other than crate::ConfigOption::Slabs, crate::ConfigOption::MaxEntries or crate::ConfigOption::MaxLoadFactor is specified, if the slab_allocator’s slab_size is greater than 65,536, or slab_count is greater than 281_474_976_710_655, max_entries is 0 or max_load_factor is not greater than 0 and less than or equal to 1.

Examples

 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
         // create a default slab allocator
         let slabs = slab_allocator!()?;

         // create a hashtable with the specified parameters
         let mut hashtable = hashtable!(MaxEntries(1_000), MaxLoadFactor(0.9), Slabs(&slabs))?;

         // do an insert, rust will figure out what type is being inserted
         hashtable.insert(&1, &2)?;

         // assert that the entry was inserted
         assert_eq!(hashtable.get(&1)?, Some(2));

         // create another hashtable with defaults, this time the global slab allocator will be
         // used. Since we did not initialize it default values will be used.
         let mut hashtable = hashtable!()?;

         // do an insert, rust will figure out what type is being inserted
         hashtable.insert(&1, &3)?;

         // assert that the entry was inserted
         assert_eq!(hashtable.get(&1)?, Some(3));

         Ok(())
 }