Macro bmw_util::hashtable_sync

macro_rules! hashtable_sync {
    ($($config:tt)*) => { ... };
}

The difference between this macro and the crate::hashtable macro is that the returned crate::Hashtable implements the Send and Sync traits and is thread safe. With this hashtable you cannot specify a crate::SlabAllocator because they use std::cell::RefCell which is not thread safe. That is also why this macro returns an error if [crate::ConfigOption::Slabs] is specified. The parameters for this macro are:

MaxEntries(usize) (optional) - The maximum number of entries that can be in this hashtable at any given time. If not specified, the default value of 100_000 will be used.

MaxLoadFactor(usize) (optional) - The maximum load factor of the hashtable. The hashtable is array based hashtable and it has a fixed size. Once the load factor is reach, insertions will return an error. The hashtable uses linear probing to handle collisions. The max_load_factor makes sure no additional insertions occur at a given ratio of entries to capacity in the array. Note that MaxEntries can always be inserted, it’s the capacity of the array that becomes larger as this ratio goes down. If not specified, the default value is 0.8.

SlabSize(usize) (optional) - the size in bytes of the slabs for this slab allocator. if not specified, the default value of 256 is used.

SlabCount(usize) (optional) - the number of slabs to allocate to this slab allocator. If not specified, the default value of 40,960 is used.

See the crate for examples.