Macro bmw_util::pattern

macro_rules! pattern {
    ( $( $pattern_items:tt)* ) => { ... };
}

The pattern macro builds a crate::Pattern which is used by the [crate::SuffixTree]. The pattern macro takes the following parameters:

Regex(String) (required) - The regular expression to use for matching (note this is not a full regular expression. Only some parts of regular expressions are implemented like wildcards and carets). See crate::Pattern for full details.

Id(usize) (required) - The id for this pattern. This id is returned in the crate::Match array if this match occurs when the [crate::SuffixTree::match] function is called.

IsMulti(bool) (optional) - If true is specified this pattern is a multi-line pattern meaning that wildcards can cross newlines. Otherwise newlines are not allowed in wildcard matches.

IsTerm(bool) (optional) - If true, this is a termination pattern meaning that if it is found, when the [crate::SuffixTree::match] function is called, matching will terminate and the matches found up to that point in the text will be returned.

IsCaseSensitive(bool) (optional) - If true only case sensitive matches are returned for this pattern. Otherwise, case-insensitive matches are also returned.

Return

Returns Ok(Pattern) on success and on error a [bmw_err::Error] is returned.

Errors

[bmw_err::ErrorKind::Configuration] - If a Regex or Id is not specified.

Examples

See crate::search_trie! for examples.