Macro bmw_util::queue

macro_rules! queue {
    ( $size:expr, $default:expr ) => { ... };
}

The crate::queue macro creates a crate::Queue implementation with the specified parameters. Note that like crate::Stack, this is a bounded queue. The queue is returned as an impl Queue<T>.

Input Parameters

size (usize) (required) - the size of the underlying array

default (bmw_ser::Serializable) (required) - a reference to the value to initialize the array with for the queue, these values are never used, but a default is needed to initialize the underlying array.

Return

Returns Ok(impl Queue<T>) on success and a bmw_err::Error on failure.

Errors

bmw_err::ErrKind::IllegalArgument - if the size is 0.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::*;

 fn main() -> Result<(), Error> {
         let mut queue = queue!(10, &0)?;

         for i in 0..10 {
                 queue.enqueue(i)?;
         }

         for i in 0..10 {
                 let v = queue.dequeue().unwrap();
                 assert_eq!(v, &i);
         }
         
         Ok(())
 }