Macro bmw_util::slab_allocator

macro_rules! slab_allocator {
    ( $( $config:expr ),* ) => { ... };
}

The slab_allocator macro initializes a slab allocator with the specified parameters. It takes the following parameters:

SlabSize(usize) (optional) - the size in bytes of the slabs for this slab allocator. if not specified, the default value of 256 is used.

SlabCount(usize) (optional) - the number of slabs to allocate to this slab allocator. If not specified, the default value of 40,960 is used.

Return

Return Ok(Rc<RefCell<dyn SlabAllocator>>) on success or bmw_err::Error on failure.

Errors

bmw_err::ErrorKind::Configuration - Is returned if a crate::ConfigOption other than crate::ConfigOption::SlabSize or crate::ConfigOption::SlabCount is specified.

bmw_err::ErrorKind::IllegalArgument - Is returned if the SlabSize is 0 or the SlabCount is 0.

Examples

 use bmw_util::*;
 use bmw_err::Error;

 fn main() -> Result<(), Error> {
     let slabs = slab_allocator!(SlabSize(128), SlabCount(5000))?;

     // this will use the specified slab allocator
     let hashtable: Box<dyn Hashtable<u32, u32>> = hashtable_box!(Slabs(&slabs))?;

     // this will also use the specified slab allocator
     // (they may be shared within the thread)
     let hashtable2: Box<dyn Hashtable<u32, u32>> = hashtable_box!(
             Slabs(&slabs),
             MaxEntries(1_000),
             MaxLoadFactor(0.9)
     )?;

     // ...

     Ok(())
 }