Struct bmw_util::Builder

pub struct Builder {}

The builder struct for the util library. All data structures are built through this interface. This is used by the macros as well.

Implementations

impl Builder

The crate::Builder is used to build the data structures in the crate.

pub fn build_thread_pool<T, OnPanic>(
config: ThreadPoolConfig
) -> Result<impl ThreadPool<T, OnPanic>, Error> where
OnPanic: FnMut(u128, Box<dyn Any + Send>) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin,
T: 'static + Send + Sync,

Build a crate::ThreadPool based on the specified crate::ThreadPoolConfig. The crate::ThreadPool::start function must be called before executing tasks.

pub fn build_array<T>(size: usize, default: &T) -> Result<Array<T>, Error> where
T: Clone,

Build a crate::Array. size is the size of the array and default is the value that is used to initialize the array. Because default must be cloned, the std::clone::Clone trait must be implemented by T. The returned value is either a crate::Array or a bmw_err::Error.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_array_list<T>(
size: usize,
default: &T
) -> Result<impl SortableList<T>, Error> where
T: Clone + Debug + PartialEq + Serializable,

Build an crate::ArrayList based on the specified size and default value. The default value is only used to initialize the underlying crate::Array and is not included in the list. On success an anonymous impl of crate::SortableList is returned.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_array_list_box<T>(
size: usize,
default: &T
) -> Result<Box<dyn SortableList<T>>, Error> where
T: Clone + Debug + PartialEq + Serializable + 'static,

boxed version of crate::Builder::build_array_list.

pub fn build_array_list_sync<T>(
size: usize,
default: &T
) -> Result<impl SortableList<T> + Send + Sync, Error> where
T: Clone + Debug + PartialEq + Serializable + Send + Sync,

sync version of crate::Builder::build_array_list.

pub fn build_array_list_sync_box<T>(
size: usize,
default: &T
) -> Result<Box<dyn SortableList<T> + Send + Sync>, Error> where
T: Clone + Debug + PartialEq + Serializable + Send + Sync + 'static,

sync box version of crate::Builder::build_array_list.

pub fn build_queue<T>(size: usize, default: &T) -> Result<impl Queue<T>, Error> where
T: Clone,

Build an crate::Queue based on the specified size and default value. The default value is only used to initialize the underlying crate::Array and is not included in the queue. On success an anonymous impl of crate::Queue is returned.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_queue_box<T>(
size: usize,
default: &T
) -> Result<Box<dyn Queue<T>>, Error> where
T: Clone + 'static,

Build an crate::Queue based on the specified size and default value. The default value is only used to initialize the underlying crate::Array and is not included in the queue. On success a Box<dyn Queue> is returned. This function may be used if you wish to store the list in a struct or enum.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_queue_sync<T>(
size: usize,
default: &T
) -> Result<impl Queue<T> + Send + Sync, Error> where
T: Clone + Send + Sync,

Build an crate::Queue based on the specified size and default value. The default value is only used to initialize the underlying crate::Array and is not included in the queue. On success an anonymous impl of crate::Queue is returned. This version requires that T be Send and Sync and returns a “Send/Sync” queue.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_queue_sync_box<T>(
size: usize,
default: &T
) -> Result<Box<dyn Queue<T> + Send + Sync>, Error> where
T: Clone + Send + Sync + 'static,

Build an crate::Queue based on the specified size and default value. The default value is only used to initialize the underlying crate::Array and is not included in the queue. On success a Box<dyn Queue> is returned. This function may be used if you wish to store the list in a struct or enum. This version requires that T be Send and Sync and returns a “Send/Sync” queue.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_stack<T>(size: usize, default: &T) -> Result<impl Stack<T>, Error> where
T: Clone,

Build an crate::Stack based on the specified size and default value. The default value is only used to initialize the underlying crate::Array and is not included in the stack. On success an anonymous impl of crate::Stack is returned.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_stack_box<T>(
size: usize,
default: &T
) -> Result<Box<dyn Stack<T>>, Error> where
T: Clone + 'static,

Build a crate::Stack based on the specified size and default value. The default value is only used to initialize the underlying crate::Array and is not included in the stack. On success a Box<dyn Stack> is returned. This function may be used if you wish to store the list in a struct or enum.

Errors

bmw_err::ErrorKind::IllegalArgument is returned if the specified size is 0.

pub fn build_stack_sync<T>(
size: usize,
default: &T
) -> Result<impl Stack<T> + Send + Sync, Error> where
T: Send + Sync + Clone,

sync version of crate::Builder::build_stack.

pub fn build_stack_sync_box<T>(
size: usize,
default: &T
) -> Result<Box<dyn Stack<T> + Send + Sync>, Error> where
T: Send + Sync + Clone + 'static,

sync box version of crate::Builder::build_stack.

pub fn build_hashtable_sync<K, V>(
config: HashtableConfig,
slab_config: SlabAllocatorConfig
) -> Result<impl Hashtable<K, V> + Send + Sync, Error> where
K: Serializable + Hash + PartialEq + Debug + Clone,
V: Serializable + Clone,

Build a synchronous crate::Hashtable based on the specified config and slab_config. The returned Hashtable implements Send and Sync. Since a shared slab allocator is not thread safe and the global slab allocator is thread local, a dedicated slab allocator must be used. That is why the slab allocator configuration is specified and a slab allocator may not be passed in as is the case with the regular hashtable/hashset builder functions. The returned value is an anonymous impl of Hashtable<K, V>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_hashtable_sync_box<K, V>(
config: HashtableConfig,
slab_config: SlabAllocatorConfig
) -> Result<Box<dyn Hashtable<K, V> + Send + Sync>, Error> where
K: Serializable + Hash + PartialEq + Debug + 'static + Clone,
V: Serializable + Clone,

Build a synchronous crate::Hashtable based on the specified config and slab_config. The returned Hashtable implements Send and Sync. Since a shared slab allocator is not thread safe and the global slab allocator is thread local, a dedicated slab allocator must be used. That is why the slab allocator configuration is specified and a slab allocator may not be passed in as is the case with the regular hashtable/hashset builder functions. The returned value is a Box<dyn Hashtable<K, V>>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_hashtable<K, V>(
config: HashtableConfig,
slabs: &Option<&Rc<RefCell<dyn SlabAllocator>>>
) -> Result<impl Hashtable<K, V>, Error> where
K: Serializable + Hash + PartialEq + Debug + Clone,
V: Serializable + Clone,

Build a crate::Hashtable based on the specified config and slabs. The returned Hashtable is not thread safe and does not implement Send or Sync. The slab allocator may be shared among other data structures, but it must not be used in other threads. The returned value is an anonymous impl of Hashtable<K, V>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_hashtable_box<K, V>(
config: HashtableConfig,
slabs: &Option<&Rc<RefCell<dyn SlabAllocator>>>
) -> Result<Box<dyn Hashtable<K, V>>, Error> where
K: Serializable + Hash + PartialEq + Debug + 'static + Clone,
V: Serializable + Clone,

Build a crate::Hashtable based on the specified config and slabs. The returned Hashtable is not thread safe and does not implement Send or Sync. The slab allocator may be shared among other data structures, but it must not be used in other threads. The returned value is a Box<dyn Hashtable<K, V>>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_hashset_sync<K>(
config: HashsetConfig,
slab_config: SlabAllocatorConfig
) -> Result<impl Hashset<K> + Send + Sync, Error> where
K: Serializable + Hash + PartialEq + Debug + Clone,

Build a synchronous crate::Hashset based on the specified config and slab_config. The returned Hashset implements Send and Sync. Since a shared slab allocator is not thread safe and the global slab allocator is thread local, a dedicated slab allocator must be used. That is why the slab allocator configuration is specified and a slab allocator may not be passed in as is the case with the regular hashtable/hashset builder functions. The returned value is an anonymous impl of Hashset<K, V>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_hashset_sync_box<K>(
config: HashsetConfig,
slab_config: SlabAllocatorConfig
) -> Result<Box<dyn Hashset<K> + Send + Sync>, Error> where
K: Serializable + Hash + PartialEq + Debug + 'static + Clone,

Build a synchronous crate::Hashset based on the specified config and slab_config. The returned Hashset implements Send and Sync. Since a shared slab allocator is not thread safe and the global slab allocator is thread local, a dedicated slab allocator must be used. That is why the slab allocator configuration is specified and a slab allocator may not be passed in as is the case with the regular hashtable/hashset builder functions. The returned value is a Box<dyn Hashset>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_hashset<K>(
config: HashsetConfig,
slabs: &Option<&Rc<RefCell<dyn SlabAllocator>>>
) -> Result<impl Hashset<K>, Error> where
K: Serializable + Hash + PartialEq + Debug + Clone,

Build a crate::Hashset based on the specified config and slabs. The returned Hashset is not thread safe and does not implement Send or Sync. The slab allocator may be shared among other data structures, but it must not be used in other threads. The returned value is an anonymous impl of Hashset<K, V>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_hashset_box<K>(
config: HashsetConfig,
slabs: &Option<&Rc<RefCell<dyn SlabAllocator>>>
) -> Result<Box<dyn Hashset<K>>, Error> where
K: Serializable + Hash + PartialEq + Debug + 'static + Clone,

Build a crate::Hashset based on the specified config and slabs. The returned Hashset is not thread safe and does not implement Send or Sync. The slab allocator may be shared among other data structures, but it must not be used in other threads. The returned value is a Box<dyn Hashset>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_list_sync<V>(
config: ListConfig,
slab_config: SlabAllocatorConfig
) -> Result<impl SortableList<V>, Error> where
V: Serializable + Debug + PartialEq + Clone,

Build a synchronous crate::List based on the specified config and slab_config. The returned List implements Send and Sync. Since a shared slab allocator is not thread safe and the global slab allocator is thread local, a dedicated slab allocator must be used. That is why the slab allocator configuration is specified and a slab allocator may not be passed in as is the case with the regular list builder functions. The returned value is an anonymous impl of SortableList. This version of the list is a linked list.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, or the slab_size is to small to fit the pointer values needed.

pub fn build_list_sync_box<V>(
config: ListConfig,
slab_config: SlabAllocatorConfig
) -> Result<Box<dyn SortableList<V> + Send + Sync>, Error> where
V: Serializable + Debug + PartialEq + Clone + 'static,

Build a synchronous crate::List based on the specified config and slab_config. The returned List implements Send and Sync. Since a shared slab allocator is not thread safe and the global slab allocator is thread local, a dedicated slab allocator must be used. That is why the slab allocator configuration is specified and a slab allocator may not be passed in as is the case with the regular list builder functions. The returned value is a Box<dyn SortableList>. This version of the list is a linked list.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, or the slab_size is to small to fit the pointer values needed.

pub fn build_list<V>(
config: ListConfig,
slabs: &Option<&Rc<RefCell<dyn SlabAllocator>>>
) -> Result<impl SortableList<V>, Error> where
V: Serializable + Debug + Clone,

Build a crate::List based on the specified config and slabs. The returned List is not thread safe and does not implement Send or Sync. The slab allocator may be shared among other data structures, but it must not be used in other threads. The returned value is an anonymous impl of the Sortable List trait.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_list_box<V>(
config: ListConfig,
slabs: &Option<&Rc<RefCell<dyn SlabAllocator>>>
) -> Result<Box<dyn SortableList<V>>, Error> where
V: Serializable + Debug + PartialEq + Clone + 'static,

Build a crate::List based on the specified config and slabs. The returned List is not thread safe and does not implement Send or Sync. The slab allocator may be shared among other data structures, but it must not be used in other threads. The returned value is a Box<dyn SortableList>.

Errors

bmw_err::ErrorKind::Configuration is returned if the slab_size is greater than 65_536, the slab count is greater than 281_474_976_710_655, max_entries is equal to 0, max_load_factor is 0 or less or greater than 1.0. or the slab_size is to small to fit the pointer values needed.

pub fn build_match(start: usize, end: usize, id: usize) -> Match

Build a match with the specified start, end and id values.

pub fn build_match_default() -> Match

Build a default match struct.

pub fn build_pattern(
regex: &str,
is_case_sensitive: bool,
is_termination_pattern: bool,
is_multi_line: bool,
id: usize
) -> Pattern

Build a pattern based on the specified regex. If is_case_sensitive is true, only case sensitive matches will be returned. Otherwise all case matches will be returned. If termination_pattern is true, if the suffix tree finds this pattern, it will stop searching for additional patterns. If is_multi_line is true, wildcard matches will be allowed to contain newlines. Otherwise a newline will terminate any potential wild card match. The id is a value that is returned in the matches array so indicate that this pattern was matched.

pub fn build_suffix_tree(
patterns: impl SortableList<Pattern>,
termination_length: usize,
max_wildcard_length: usize
) -> Result<impl SuffixTree + Send + Sync, Error>

Builds a suffix tree based on the specified list of patterns. The termination_length is the length at which the matching terminates. The max_wildcard_length is the maximum length of any wild card matches.

pub fn build_suffix_tree_box(
patterns: impl SortableList<Pattern>,
termination_length: usize,
max_wildcard_length: usize
) -> Result<Box<dyn SuffixTree + Send + Sync>, Error>

Same as crate::Builder::build_suffix_tree except that the tree is returned as a Box>.

pub fn build_slabs_unsafe() -> UnsafeCell<Box<dyn SlabAllocator>>

Build a slab allocator on the heap in an std::cell::UnsafeCell. This function is used by the global thread local slab allocator to allocate thread local slab allocators. Note that it calls unsafe functions. This function should generally be called through the crate::global_slab_allocator macro.

pub fn build_slabs_ref() -> Rc<RefCell<dyn SlabAllocator>>

Build a slab allocator in a Rc/RefCell. This function is used by crate::slab_allocator to create slab allocators for use with the other macros.

pub fn build_slabs() -> Box<dyn SlabAllocator>

Build a slab allocator in a Box.

pub fn build_sync_slabs() -> Box<dyn SlabAllocator + Send + Sync>

sync version of crate::Builder::build_slabs.

pub fn build_lock<T>(t: T) -> Result<impl Lock<T>, Error> where
T: Send + Sync,

Build a crate::Lock.

pub fn build_lock_box<T>(t: T) -> Result<Box<dyn LockBox<T>>, Error> where
T: Send + Sync + 'static,

Build a crate::LockBox.