Trait bmw_util::Hashset

pub trait Hashset<K>: Debug + DynClone where
    K: Serializable + Clone, {
    fn max_load_factor(&self) -> f64;
    fn max_entries(&self) -> usize;
    fn insert(&mut self, key: &K) -> Result<(), Error>;
    fn contains(&self, key: &K) -> Result<bool, Error>;
    fn remove(&mut self, key: &K) -> Result<bool, Error>;
    fn size(&self) -> usize;
    fn clear(&mut self) -> Result<(), Error>;
    fn iter<'a>(&'a self) -> HashsetIterator<'a, K>;
}

A slab allocated Hashset. Most of the implementation is shared with crate::Hashtable. See crate::Hashtable for a discussion of the slab layout. The difference is that as is the case with hashsets, there is no value.

Examples

 use bmw_err::*;
 use bmw_util::*;

 fn main() -> Result<(), Error> {
      let slabs = slab_allocator!()?;
      let mut hashset = hashset!(MaxEntries(1_000), Slabs(&slabs))?;

      hashset.insert(&1)?;
      assert_eq!(hashset.size(), 1);
      assert!(hashset.contains(&1)?);
      assert!(!hashset.contains(&2)?);
      assert_eq!(hashset.remove(&1)?, true);
      assert_eq!(hashset.remove(&1)?, false);
      assert_eq!(hashset.size(), 0);

      hashset.insert(&1)?;
      hashset.clear()?;
      assert_eq!(hashset.size(), 0);

      Ok(())
 }

Required Methods

fn max_load_factor(&self) -> f64

Returns the maximum load factor as configured for this crate::Hashset.

fn max_entries(&self) -> usize

Returns the maximum entries as configured for this crate::Hashset.

fn insert(&mut self, key: &K) -> Result<(), Error>

Insert a key into this hashset.

fn contains(&self, key: &K) -> Result<bool, Error>

If key is present this function returns true, otherwise false.

fn remove(&mut self, key: &K) -> Result<bool, Error>

Remove the specified key from this hashset.

fn size(&self) -> usize

Returns the size of this hashset.

fn clear(&mut self) -> Result<(), Error>

Clear all items, reinitialized the entry array, and free the slabs associated with this hashset.

fn iter<'a>(&'a self) -> HashsetIterator<'a, K>

Returns an std::iter::Iterator to iterate through this hashset.