Trait bmw_util::Hashtable

pub trait Hashtable<K, V>: Debug + DynClone
where
    K: Serializable + Clone,
    V: Serializable,
{

    // Required methods
    fn max_load_factor(&self) -> f64;
    fn max_entries(&self) -> usize;
    fn insert(&mut self, key: &K, value: &V) -> Result<(), Error>;
    fn get(&self, key: &K) -> Result<Option<V>, Error>;
    fn remove(&mut self, key: &K) -> Result<Option<V>, Error>;
    fn size(&self) -> usize;
    fn clear(&mut self) -> Result<(), Error>;
    fn iter<'a>(&'a self) -> HashtableIterator<'a, K, V> ;
    fn bring_to_front(&mut self, key: &K) -> Result<(), Error>;
    fn remove_oldest(&mut self) -> Result<(), Error>;
    fn raw_read(
        &self,
        key: &K,
        offset: usize,
        data: &mut [u8; 412]
    ) -> Result<bool, Error>;
    fn raw_write(
        &mut self,
        key: &K,
        offset: usize,
        data: &[u8; 412],
        len: usize
    ) -> Result<(), Error>;
    fn slabs(&self) -> Result<Option<Rc<RefCell<dyn SlabAllocator>>>, Error>;

}

A slab allocated hashtable. Data is stored in a crate::SlabAllocator defined by the user or using a global thread local slab allocator. All keys and values must implement the bmw_ser::Serializable trait which can be implemented with the bmw_derive::Serializable proc_macro.

Sizing

It is important to configure this hashtable correctly and to do so, the user must understand how the data are stored. Each hashtable is configured with an associated crate::SlabAllocator and that slab allocator has a specific size for each slab. The hashtable will only store a single entry in a slab. No slab contains data from multiple entries. So, if you have a large slab size and small entries, you will be wasting space. If your data is fixed in size, you can calculate an optimal slab size, but if it is variable, it is probably better to use smaller slabs so that less space is left empty. The slab layout looks like this:

[ ptr_size bytes for next iterator list] [ ptr_size bytes for prev iterator list] [ hashtable key ] [ hashtable value ] … (empty space if the slab is not filled up) [ ptr_size bytes for key/value list ]

where ptr_size is determined by the size of the entry array. The minimum number of bytes will be used. So, if you have an entry array that is length less than 256 ptr_size will be 1 byte. If you have an entry array that is length less than 65,536 but greater than or equal to 256, ptr_size will be 2 and so on. If the data for the entry only takes up one slab, the key/value list is not used, but if the data takes up more than one slab, the key/value list will point to the next slab in the list. So, if your key is always 8 bytes and your value is always 8 bytes and you have an entry array size of 100,000 (ptr_size = 3 bytes), you can will need a total of 9 bytes overhead for the three pointers and you will need 16 bytes for your data. So you can size your hashtable at 25 bytes.

Note: your entry array size is ceil(the max_entries / max_load_factor).

Examples

 use bmw_err::*;
 use bmw_util::*;
 use bmw_log::*;

 info!();

 fn main() -> Result<(), Error> {
         let slabs = slab_allocator!(SlabSize(128), SlabCount(10_000))?;
         let mut hashtable = hashtable!(Slabs(&slabs))?;

         hashtable.insert(&1, &2)?;
         let v = hashtable.get(&1)?.unwrap();
         info!("v={}", v)?;
         assert_eq!(v, 2);

         Ok(())
 }

 use bmw_err::*;
 use bmw_util::*;
 use bmw_log::*;
 use std::collections::HashMap;
 use bmw_deps::rand::random;

 info!();

 fn main() -> Result<(), Error> {
     let mut keys = vec![];
     let mut values = vec![];
     for _ in 0..1_000 {
         keys.push(random::<u32>());
         values.push(random::<u32>());
     }
     let mut hashtable = Builder::build_hashtable(HashtableConfig::default(), &None)?;
     let mut hashmap = HashMap::new();
     for i in 0..1_000 {
         hashtable.insert(&keys[i], &values[i])?;
         hashmap.insert(&keys[i], &values[i]);
     }

     for _ in 0..100 {
         let index: usize = random::<usize>() % 1_000;
         hashtable.remove(&keys[index])?;
         hashmap.remove(&keys[index]);
     }

     let mut i = 0;
     for (k, vm) in &hashmap {
         let vt = hashtable.get(&k)?;
         assert_eq!(&vt.unwrap(), *vm);
         i += 1;
     }

     assert_eq!(i, hashtable.size());
     assert_eq!(i, hashmap.len());

     Ok(())
 }

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;
 use bmw_derive::*;

 info!();

 #[derive(Serializable, Clone, Debug, PartialEq)]
 struct MyStruct {
     id: u128,
     name: String,
     phone: Option<String>,
     age: u8,
 }

 fn main() -> Result<(), Error> {
     let s = MyStruct {
         id: 1234,
         name: "Hagrid".to_string(),
         phone: None,
         age: 54,
     };

     debug!("my struct = {:?}", s)?;

     let mut hashtable = hashtable!()?;

     hashtable.insert(&1, &s)?;

     let v = hashtable.get(&1)?;
     assert_eq!(v, Some(s));

     info!("value of record #1 is {:?}", v)?;

     Ok(())
 }

Required Methods

fn max_load_factor(&self) -> f64

Returns the maximum load factor as configured for this crate::Hashtable.

fn max_entries(&self) -> usize

Returns the maximum entries as configured for this crate::Hashtable.

fn insert(&mut self, key: &K, value: &V) -> Result<(), Error>

Insert a key/value pair into the hashtable.

fn get(&self, key: &K) -> Result<Option<V>, Error>

Get the value associated with the specified key.

fn remove(&mut self, key: &K) -> Result<Option<V>, Error>

Remove the specified key from the hashtable.

fn size(&self) -> usize

Return the size of the hashtable.

fn clear(&mut self) -> Result<(), Error>

Clear all items, reinitialized the entry array, and free the slabs associated with this hashtable.

fn iter<'a>(&'a self) -> HashtableIterator<'a, K, V>

Returns an std::iter::Iterator to iterate through this hashtable.

fn bring_to_front(&mut self, key: &K) -> Result<(), Error>

Bring the entry to the front of the list for deletion purposes in a cache.

fn remove_oldest(&mut self) -> Result<(), Error>

Remove the oldest entry in the hashtable.

fn raw_read( &self, key: &K, offset: usize, data: &mut [u8; 412] ) -> Result<bool, Error>

Get raw data and store it in data with given offset.

fn raw_write( &mut self, key: &K, offset: usize, data: &[u8; 412], len: usize ) -> Result<(), Error>

Write raw data from data with given offset.

fn slabs(&self) -> Result<Option<Rc<RefCell<dyn SlabAllocator>>>, Error>

Gets the slab allocator associated with this Hashtable or None if the global slab allocator is used.