Trait bmw_util::List

pub trait List<V>: DynClone + Debug {
    fn push(&mut self, value: V) -> Result<(), Error>;
    fn iter<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a>
    where
        V: Serializable + Clone;
    fn iter_rev<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a>
    where
        V: Serializable + Clone;
    fn delete_head(&mut self) -> Result<(), Error>;
    fn size(&self) -> usize;
    fn clear(&mut self) -> Result<(), Error>;
}

A trait that defines a list. Both an array list and a linked list are implemented the default crate::list macro returns the linked list. The crate::array_list macro returns the array list. Both implement this trait.

Examples

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
     // for this example we will use the global slab allocator which is a
     // thread local slab allocator which we can configure via macro
     global_slab_allocator!(SlabSize(64), SlabCount(100_000))?;

     // create two lists (one linked and one array list).
     // Note that all lists created via macro are interoperable.
     let mut list1 = list![1u32, 2u32, 4u32, 5u32];
     let mut list2 = array_list!(10, &0u32)?;
     list2.push(5)?;
     list2.push(2)?;
     list2.push(4)?;
     list2.push(1)?;

     // sort the array list and assert it's equal to the other list
     list2.sort()?;
     assert!(list_eq!(list1, list2));

     // append list2 to list1 and assert the value
     list_append!(list1, list2);
     assert!(list_eq!(list1, list![1, 2, 4, 5, 1, 2, 4, 5]));

     Ok(())
 }

Required Methods

fn push(&mut self, value: V) -> Result<(), Error>

push a value onto the end of the list

fn iter<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a> where
V: Serializable + Clone,

iterate through the list

fn iter_rev<'a>(&'a self) -> Box<dyn Iterator<Item = V> + 'a> where
V: Serializable + Clone,

iterate through the list in reverse order

fn delete_head(&mut self) -> Result<(), Error>

delete the head of the list

fn size(&self) -> usize

return the size of the list

fn clear(&mut self) -> Result<(), Error>

clear all items from the list