Trait bmw_util::LockBox

pub trait LockBox<T>: Send + Sync + Debug where
    T: Send + Sync, {
    fn wlock(&mut self) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>;
    fn rlock(&self) -> Result<RwLockReadGuardWrapper<'_, T>, Error>;
    fn wlock_ignore_poison(
        &mut self
    ) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>;
    fn danger_to_usize(&self) -> usize;
    fn inner(&self) -> Arc<RwLock<T>>;
    fn id(&self) -> u128;
}

crate::LockBox is the same as crate::Lock except that it is possible to build The LockBox into a Box<dyn LockBox> structure so that it is object safe. It can then be cloned using DynClone.

Examples

 use bmw_err::*;
 use bmw_util::*;

 struct TestLockBox<T> {
     lock_box: Box<dyn LockBox<T>>,
 }

 fn test_lock_box() -> Result<(), Error> {
     let lock_box = lock_box!(1u32)?;
     let mut lock_box2 = lock_box.clone();
     let mut tlb = TestLockBox { lock_box };

     {
         let mut tlb = tlb.lock_box.wlock()?;
         (**tlb.guard()) = 2u32;
     }

     {
         let tlb = tlb.lock_box.rlock()?;
         assert_eq!((**tlb.guard()), 2u32);
     }

     Ok(())
 }

Required Methods

fn wlock(&mut self) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>

obtain a write lock and corresponding std::sync::RwLockWriteGuard for this crate::LockBox.

fn rlock(&self) -> Result<RwLockReadGuardWrapper<'_, T>, Error>

obtain a read lock and corresponding std::sync::RwLockReadGuard for this crate::LockBox.

fn wlock_ignore_poison(
&mut self
) -> Result<RwLockWriteGuardWrapper<'_, T>, Error>

Same as crate::LockBox::wlock except that any poison errors are ignored by calling the underlying into_inner() fn.

fn danger_to_usize(&self) -> usize

consume the inner Arc and return a usize value. This function is dangerous because it potentially leaks memory. The usize must be rebuilt into a lockbox that can then be dropped via the crate::lock_box_from_usize function.

fn inner(&self) -> Arc<RwLock<T>>

return the inner data holder.

fn id(&self) -> u128

return the id for this lockbox.