Trait bmw_util::Queue

pub trait Queue<V>: DynClone {
    // Required methods
    fn enqueue(&mut self, value: V) -> Result<(), Error>;
    fn dequeue(&mut self) -> Option<&V>;
    fn peek(&self) -> Option<&V>;
    fn length(&self) -> usize;
}

This trait defines a queue. The implementation is a bounded queue. The queue uses an crate::Array as the underlying storage mechanism.

Examples:

 // import the util/log/err libraries
 use bmw_util::*;
 use bmw_log::*;
 use bmw_err::*;

 fn main() -> Result<(), Error> {
     // create a queue with capacity of 1_000 items and &0
     //is the default value used to initialize the array
     let mut queue = queue!(1_000, &0)?;

     // add three items to the queue
     queue.enqueue(1)?;
     queue.enqueue(2)?;
     queue.enqueue(3)?;

     // dequeue the first item
     assert_eq!(queue.dequeue(), Some(&1));

     // dequeue the second item
     assert_eq!(queue.dequeue(), Some(&2));

     // peek at the last item
     assert_eq!(queue.peek(), Some(&3));

     // dequeue the last item
     assert_eq!(queue.dequeue(), Some(&3));

     // now that the queue is empty, dequeue returns None
     assert_eq!(queue.dequeue(), None);

     Ok(())
 }

Required Methods

fn enqueue(&mut self, value: V) -> Result<(), Error>

Enqueue a value

fn dequeue(&mut self) -> Option<&V>

Dequeue a value. If the queue is Empty return None

fn peek(&self) -> Option<&V>

peek at the next value in the queue. If the queue is Empty return None

fn length(&self) -> usize

return the number of items currently in the queue