Trait bmw_util::SlabAllocator

pub trait SlabAllocator: DynClone + Debug {
    // Required methods
    fn is_init(&self) -> bool;
    fn allocate<'a>(&'a mut self) -> Result<SlabMut<'a>, Error>;
    fn free(&mut self, id: usize) -> Result<(), Error>;
    fn get<'a>(&'a self, id: usize) -> Result<Slab<'a>, Error>;
    fn get_mut<'a>(&'a mut self, id: usize) -> Result<SlabMut<'a>, Error>;
    fn free_count(&self) -> Result<usize, Error>;
    fn slab_size(&self) -> Result<usize, Error>;
    fn slab_count(&self) -> Result<usize, Error>;
    fn init(&mut self, config: SlabAllocatorConfig) -> Result<(), Error>;
}

This trait defines the public interface to the crate::SlabAllocator. The slab allocator is used by the other data structures in this crate to avoid dynamic heap allocations. By itself, the slab allocator is fairly simple. It only allocates and frees slabs. crate::SlabAllocator::get and crate::SlabAllocator::get_mut are also provided to obtain immutable and mutable references to a slab respectively. They only contain references to the data and not copies.

Examples

 use bmw_err::*;
 use bmw_util::slab_allocator;
 use std::cell::{RefMut,Ref};

 fn main() -> Result<(), Error> {
     // build a slab allocator, in this case with defaults
     let slabs = slab_allocator!()?;

     let id = {
         // slab allocator is stored in an Rc<RefCell>. This allows for it to be used by
         // multiple data structures at the same time.
         let mut slabs: RefMut<_> = slabs.borrow_mut();
         // allocate a slab. [`crate::SlabAllocator::allocate`] returns [`crate::SlabMut`]
         // which contains a mutable reference to the underlying data in the slab.
         let mut slab = slabs.allocate()?;

         // get the id for this slab
         let id = slab.id();
         // get_mut returns a mutable reference to the data in owned by the
         // [`crate::SlabAllocator`]
         slab.get_mut()[0] = 101;
         id
     };

     // borrow, this time with a Ref instead of refmut since it's an immutable call.
     let slabs: Ref<_> = slabs.borrow();
     // now we can get an immutable reference to this slab
     let slab = slabs.get(id)?;
     assert_eq!(slab.get()[0], 101);

     Ok(())
 }

Required Methods

fn is_init(&self) -> bool

If the slab allocator has been initialized, return true, otherwise, false.

fn allocate<'a>(&'a mut self) -> Result<SlabMut<'a>, Error>

Allocate a slab and return a crate::SlabMut on success. On failure, return an [bmw_err::Error].

[bmw_err::ErrorKind::IllegalState] if the crate::SlabAllocator::init function has not been called.

[bmw_err::ErrorKind::CapacityExceeded] if the capacity of this crate::SlabAllocator has been exceeded.

fn free(&mut self, id: usize) -> Result<(), Error>

Free a slab that has previously been allocated by this slab allocator. id is the id of the slab to free. It can be obtained through the crate::SlabMut::id or crate::Slab::id function. Return a [bmw_err::Error] on failure.

[bmw_err::ErrorKind::ArrayIndexOutOfBounds] if this slab entry is too big for this instance.

[bmw_err::ErrorKind::IllegalState] if the crate::SlabAllocator::init function has not been called or this slab was not allocated.

Examples

 use bmw_err::*;
 use bmw_util::slab_allocator;
 use bmw_util::ConfigOption::{SlabSize, SlabCount};

 fn main() -> Result<(), Error> {
     // instantiate a slab allocator with a slab count of 1,000.
     let mut slabs = slab_allocator!(SlabSize(1_000), SlabCount(1_000))?;

     // borrow a mutable reference
     let mut slabs = slabs.borrow_mut();

     // assert that there are 1,000 free slabs.
     assert_eq!(slabs.free_count()?, 1_000);

     let slab_id = {
         // allocate a slab.
         let slab = slabs.allocate()?;
         slab.id()
     };

     // assert that the free count has decreased by 1.
     assert_eq!(slabs.free_count()?, 999);

     // free the slab that was allocated
     slabs.free(slab_id)?;

     // assert that the free count has returned to the initial value of 1,000.
     assert_eq!(slabs.free_count()?, 1_000);

     Ok(())
 }

fn get<'a>(&'a self, id: usize) -> Result<Slab<'a>, Error>

Get an immutable reference to a slab that has previously been allocated by the crate::SlabAllocator. On success a crate::Slab is returned. On failure, a [bmw_err::Error] is returned.

[bmw_err::ErrorKind::ArrayIndexOutOfBounds] if this slab entry is too big for this instance.

[bmw_err::ErrorKind::IllegalState] if the crate::SlabAllocator::init function has not been called or this slab was not allocated.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::slab_allocator;
 use bmw_util::ConfigOption::{SlabSize,SlabCount};

 info!();

 fn main() -> Result<(), Error> {
     // instantiate a slab allocator with a slab count of 1,000.
     let mut slabs = slab_allocator!(SlabCount(1_000), SlabSize(1_000))?;
     
     // borrow a mutable reference
     let mut slabs = slabs.borrow_mut();

     // assert that there are 1,000 free slabs.
     assert_eq!(slabs.free_count()?, 1_000);

     let slab_id = {
         // allocate a slab.
         let slab = slabs.allocate()?;
         slab.id()
     };

     // assert that the free count has decreased by 1.
     assert_eq!(slabs.free_count()?, 999);

     // get the slab that was allocated
     let slab = slabs.get(slab_id)?;

     info!("slab data = {:?}", slab.get())?;

     Ok(())
 }

fn get_mut<'a>(&'a mut self, id: usize) -> Result<SlabMut<'a>, Error>

Get an mutable reference to a slab that has previously been allocated by the crate::SlabAllocator. On success a crate::SlabMut is returned. On failure, a [bmw_err::Error] is returned.

[bmw_err::ErrorKind::ArrayIndexOutOfBounds] if this slab entry is too big for this instance.

[bmw_err::ErrorKind::IllegalState] if the crate::SlabAllocator::init function has not been called or this slab was not allocated.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::slab_allocator;
 use bmw_util::ConfigOption::{SlabSize, SlabCount};

 info!();

 fn main() -> Result<(), Error> {
     // instantiate a slab allocator with a slab count of 1,000.
     let mut slabs = slab_allocator!(SlabSize(1_000), SlabCount(1_000))?;

     // borrow a mutable reference
     let mut slabs = slabs.borrow_mut();

     // assert that there are 1,000 free slabs.
     assert_eq!(slabs.free_count()?, 1_000);

     let slab_id = {
         // allocate a slab.
         let slab = slabs.allocate()?;
         slab.id()
     };

     // assert that the free count has decreased by 1.
     assert_eq!(slabs.free_count()?, 999);

     // get the slab that was allocated
     let mut slab = slabs.get_mut(slab_id)?;

     info!("slab data = {:?}", slab.get_mut())?;

     Ok(())
 }

fn free_count(&self) -> Result<usize, Error>

Returns the number of free slabs this crate::SlabAllocator has remaining.

fn slab_size(&self) -> Result<usize, Error>

Returns the configured slab_size for this crate::SlabAllocator.

fn slab_count(&self) -> Result<usize, Error>

Returns the configured slab_count for this crate::SlabAllocator.

fn init(&mut self, config: SlabAllocatorConfig) -> Result<(), Error>

Initializes the crate::SlabAllocator with the given config. See crate::SlabAllocatorConfig for further details.