Trait bmw_util::SortableList

pub trait SortableList<V>: List<V> + DynClone {
    // Required methods
    fn sort(&mut self) -> Result<(), Error>
       where V: Ord;
    fn sort_unstable(&mut self) -> Result<(), Error>
       where V: Ord;
}

A trait that defines a sortable list. Both implementations implement this trait, although the linked list merely copies the data into an array list to sort. The array_list can natively sort with rust’s sort implementations using the functions associated with slice.

Required Methods

fn sort(&mut self) -> Result<(), Error>
where V: Ord,

sort with a stable sorting algorithm

fn sort_unstable(&mut self) -> Result<(), Error>
where V: Ord,

sort with an unstable sorting algorithm. unstable sort is significantly faster and should be used when stable sorting (ording of equal values consistent) is not required.