Trait bmw_util::ThreadPool

pub trait ThreadPool<T, OnPanic> where
    OnPanic: FnMut(u128, Box<dyn Any + Send>) -> Result<(), Error> + Send + 'static + Clone + Sync + Unpin, {
    fn execute<F>(
        &self,
        f: F,
        id: u128
    ) -> Result<Receiver<PoolResult<T, Error>>, Error>
    where
        F: Future<Output = Result<T, Error>> + Send + 'static;
    fn start(&mut self) -> Result<(), Error>;
    fn stop(&mut self) -> Result<(), Error>;
    fn size(&self) -> Result<usize, Error>;
    fn stopper(&self) -> Result<ThreadPoolStopper, Error>;
    fn executor(&self) -> Result<ThreadPoolExecutor<T>, Error>
    where
        T: Send + Sync;
    fn set_on_panic(&mut self, on_panic: OnPanic) -> Result<(), Error>;
}

This trait defines the public interface to the ThreadPool. A pool can be configured via the crate::ThreadPoolConfig struct. The thread pool should be accessed through the macros under normal circumstances. See crate::thread_pool, crate::execute and crate::block_on for additional details. The thread pool can be passed through threads via a crate::Lock or crate::LockBox so a single thread pool can service multiple worker threads. See examples below.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_util::*;
 use bmw_util::ConfigOption::{MaxSize, MinSize};

 info!();

 fn thread_pool() -> Result<(), Error> {
     let mut tp = thread_pool!()?; // create a thread pool using default settings
     tp.set_on_panic(move |_id,_e| -> Result<(), Error> { Ok(()) })?;

     // create a shared variable protected by the [`crate::lock`] macro.
     let mut shared = lock!(0)?; // we use an integer 0, but any struct can be used.
     let shared_clone = shared.clone();

     // execute a task
     let handle = execute!(tp, {
         // obtain the lock in write mode
         let mut shared = shared.wlock()?;
         // increment the counter
         (**shared.guard()) += 1;
         // return the counter value (now it's 1)
         Ok((**shared.guard()))
     })?;

     // block on the returned handle until it's task is complete and assert that the
     // return value is 1.
     assert_eq!(block_on!(handle), PoolResult::Ok(1));

     // check our shared value which was updated in the thread pool.
     let shared_clone = shared_clone.rlock()?;
     assert_eq!((**shared_clone.guard()), 1);

     Ok(())
 }

 fn thread_pool2() -> Result<(), Error> {
     // create a thread pool with the specified max/min size. See [`crate::ThreadPoolConfig`]
     // for further details.
     let mut tp = thread_pool!(MaxSize(10), MinSize(5))?;
     tp.set_on_panic(move |_id,_e| -> Result<(), Error> { Ok(()) })?;

     // put the thread pool in a [`crate::Lock`].
     let tp = lock!(tp)?;

     // spawn 6 worker threads
     for _ in 0..6 {
         // create a clone of our locked thread pool for each worker thread
         let tp = tp.clone();
         std::thread::spawn(move || -> Result<(), Error> {
             // do some work here and pass other work to the thread pool.
             // ...

             // obtain a read lock on the thread pool
             let tp = tp.rlock()?;

             // execute the task in the thread pool, the returned handle can be used
             // if desired or ignored
             execute!((**tp.guard()), {
                 info!("executing in thread pool")?;
                 Ok(1)
             })?;
             Ok(())
         });
     }
     Ok(())
 }

Required Methods

fn execute<F>(
&self,
f: F,
id: u128
) -> Result<Receiver<PoolResult<T, Error>>, Error> where
F: Future<Output = Result<T, Error>> + Send + 'static,

Execute a task in the thread pool. This task will run to completion on the first available thread in the pool. The return value is a receiver which will be sent a crate::PoolResult on completion of the task. If an error occurs, bmw_err::Error will be returned.

fn start(&mut self) -> Result<(), Error>

Start the pool. If macros are used, this call is unnecessary.

fn stop(&mut self) -> Result<(), Error>

Stop the thread pool. This function will ensure no new tasks are processed in the ThreadPool and that the threads will be stopped once they become idle again. It however, does not ensure that any tasks currently running in the thread pool are stopped immediately. That is the responsibility of the user.

fn size(&self) -> Result<usize, Error>

Returns the current size of the thread pool which will be between crate::ThreadPoolConfig::min_size and crate::ThreadPoolConfig::max_size.

fn stopper(&self) -> Result<ThreadPoolStopper, Error>

Get the crate::ThreadPoolStopper for this thread pool.

fn executor(&self) -> Result<ThreadPoolExecutor<T>, Error> where
T: Send + Sync,

Get the crate::ThreadPoolExecutor for this thread pool.

fn set_on_panic(&mut self, on_panic: OnPanic) -> Result<(), Error>

Set an on panic handler for this thread pool